MathOptInterface.initialize — Function

initialize(
    d::AbstractNLPEvaluator,
    requested_features::Vector{Symbol},
)::Nothing

Initialize d with the set of features in requested_features. Check features_available before calling initialize to see what features are supported by d.

Warning

This method must be called before any other methods.

Features

The following features are defined:

:Grad: enables eval_objective_gradient

:Jac: enables eval_constraint_jacobian

:JacVec: enables eval_constraint_jacobian_product and eval_constraint_jacobian_transpose_product

:Hess: enables eval_hessian_lagrangian

:HessVec: enables eval_hessian_lagrangian_product

:ExprGraph: enables objective_expr and constraint_expr.

In all cases, including when requested_features is empty, eval_objective and eval_constraint are supported.

Examples

MOI.initialize(d, Symbol[])
MOI.initialize(d, [:ExprGraph])
MOI.initialize(d, MOI.features_available(d))