Reference

API

[Some introduction to API. List basic standalone methods.]

Attributes

List of attribute categories.

MathOptInterface.AbstractSolverAttribute — Type.

AbstractSolverAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of the solver.

source

MathOptInterface.AbstractSolverInstanceAttribute — Type.

AbstractSolverInstanceAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of the solver instance.

source

MathOptInterface.AbstractVariableAttribute — Type.

AbstractVariableAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of variables in the solver instance.

source

MathOptInterface.AbstractConstraintAttribute — Type.

AbstractConstraintAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of constraints in the solver instance.

source

Functions for getting and setting attributes.

MathOptInterface.cangetattribute — Function.

cangetattribute(s::AbstractSolver, attr::AbstractSolverAttribute)::Bool

Return a Bool indicating whether it is possible to query attribute attr from the solver s.

cangetattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, R::Type{VariableReference})::Bool
cangetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, R::Type{VariablewiseConstraintReference{T})::Bool
cangetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, R::Type{AffineConstraintReference{T})::Bool
cangetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, R::Type{QuadraticConstraintReference{T})::Bool

Return a Bool indicating whether the solver instance m currently has a value for the attributed specified by attribute type attr applied to the reference type R.

Examples

cangetattribute(GurobiSolver(), SupportsAffineConstraint{Zero}())
cangetattribute(m, ObjectiveValue())
cangetattribute(m, VariablePrimalStart(), varref)
cangetattribute(m, ConstraintPrimal(), conref)

source

MathOptInterface.getattribute — Function.

getattribute(s::AbstractSolver, attr::AbstractSolverAttribute)

Return an attribute attr of the solver s.

getattribute(m::AbstractSolverInstance, attr::AbstractSolverInstanceAttribute)

Return an attribute attr of the solver instance m.

getattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::VariableReference)

Return an attribute attr of the variable v in solver instance m.

getattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::Vector{VariableReference})

Return a vector of attributes corresponding to each variable in the collection v in the solver instance m.

getattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::ConstraintReference)

Return an attribute attr of the constraint c in solver instance m.

getattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{VariablewiseConstraintReference{T}})
getattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{AffineConstraintReference{T}})
getattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{QuadraticConstraintReference{T}})

Return a vector of attributes corresponding to each constraint in the collection c in the solver instance m.

Examples

getattribute(m, ObjectiveValue())
getattribute(m, VariableResult(), ref)
getattribute(m, VariableResult(5), [ref1, ref2])
getattribute(m, OtherAttribute("something specific to cplex"))

source

MathOptInterface.getattribute! — Function.

getattribute!(output, m::AbstractSolverInstance, args...)

An in-place version of getattribute. The signature matches that of getattribute except that the the result is placed in the vector output.

source

MathOptInterface.cansetattribute — Function.

cansetattribute(s::AbstractSolver, attr::AbstractSolverAttribute)::Bool

Return a Bool indicating whether it is possible to set attribute attr in the solver s.

cansetattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, R::Type{VariableReference})::Bool
cansetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, R::Type{VariablewiseConstraintReference{T})::Bool
cangetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, R::Type{AffineConstraintReference{T})::Bool
cangetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, R::Type{QuadraticConstraintReference{T})::Bool

Return a Bool indicating whether it is possible to set attribute attr applied to the reference type R in the solver instance m.

Examples

cansetattribute(GurobiSolver(), SupportsAffineConstraint{Zero}())
cansetattribute(m, ObjectiveValue())
cansetattribute(m, VariablePrimalStart(), VariableReference)
cansetattribute(m, ConstraintPrimal(), AffineConstraintReference{NonNegative})

source

MathOptInterface.setattribute! — Function.

setattribute!(s::AbstractSolver, attr::AbstractSolverAttribute, value)

Assign value to the attribute attr of the solver s.

setattribute!(m::AbstractSolverInstance, attr::AbstractSolverInstanceAttribute, value)

Assign value to the attribute attr of the solver instance m.

setattribute!(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::VariableReference, value)

Assign value to the attribute attr of variable v in solver instance m.

setattribute!(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::Vector{VariableReference}, vector_of_values)

Assign a value respectively to the attribute attr of each variable in the collection v in solver instance m.

setattribute!(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::ConstraintReference, value)

Assign a value to the attribute attr of constraint c in solver instance m.

setattribute!(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{VariablewiseConstraintReference{T}})
setattribute!(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{AffineConstraintReference{T}})
setattribute!(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{QuadraticConstraintReference{T}})

Assign a value respectively to the attribute attr of each constraint in the collection c in solver instance m.

source

Solver

MathOptInterface.AbstractSolver — Type.

AbstractSolver

Abstract supertype for "solver" objects. A solver is a lightweight object used for selecting solvers and parameters. It does not store any solver instance data.

source

List of solver attributes

MathOptInterface.ReturnsDuals — Type.

ReturnsDuals()

A Bool indicating if the solver should be expected to return dual solutions when appropriate.

source

MathOptInterface.SupportsAddConstraintAfterSolve — Type.

SupportsAddConstraintAfterSolve()

A Bool indicating if the solver supports adding constraints after a solve. If false, then a new solver instance should be constructed instead.

source

MathOptInterface.SupportsDeleteConstraint — Type.

SupportsDeleteConstraint()

A Bool indicating if the solver supports deleting constraints from a solver instance.

source

MathOptInterface.SupportsAddVariableAfterSolve — Type.

SupportsAddVariableAfterSolve()

A Bool indicating if the solver supports adding variables after a solve. In the context of linear programming, this is known as column generation.

source

MathOptInterface.SupportsQuadraticObjective — Type.

SupportsQuadraticObjective()

A Bool indicating if the solver supports quadratic objectives.

source

MathOptInterface.SupportsConicThroughQuadratic — Type.

SupportsConicThroughQuadratic()

A Bool indicating if the solver interprets certain quadratic constraints as second-order cone constraints.

source

SolverInstance

MathOptInterface.AbstractSolverInstance — Type.

AbstractSolverInstance

Abstract supertype which represents a solver's in-memory representation of an optimization problem.

source

MathOptInterface.SolverInstance — Function.

SolverInstance(solver::AbstractSolver)

Create a solver instance of AbstractSolverInstance using the given solver.

source

MathOptInterface.optimize! — Function.

optimize!(m::AbstractSolverInstance)

Start the solution procedure.

source

MathOptInterface.freesolverinstance! — Function.

freesolver instance!(m::AbstractSolverInstance)

Release any resources and memory used by the solver instance. Note that the Julia garbage collector takes care of this automatically, but automatic collection cannot always be forced. This method is useful for more precise control of resources, especially in the case of commercial solvers with licensing restrictions on the number of concurrent runs. Users must discard the solver instance object after this method is invoked.

source

List of solver instance attributes

MathOptInterface.RawSolver — Type.

RawSolver()

An object that may be used to access a solver-specific API for this solver instance.

source

MathOptInterface.Sense — Type.

Sense()

The optimization sense of the solver instance, an OptimizationSense with value MinSense or MaxSense.

source

MathOptInterface.NumberOfVariables — Type.

NumberOfVariables()

The number of variables in the solver instance.

source

MathOptInterface.NumberOfVariablewiseConstraints — Type.

NumberOfVariablewiseConstraints{T}()

The number of variablewise constraints of type T in the solver instance.

source

MathOptInterface.NumberOfAffineConstraints — Type.

NumberOfAffineConstraints{T}()

The number of affine constraints of type T in the solver instance.

source

MathOptInterface.NumberOfQuadraticConstraints — Type.

NumberOfQuadraticConstraints{T}()

The number of quadratic constraints of type T in the solver instance.

source

MathOptInterface.ResultCount — Type.

ResultCount()

The number of results available.

source

MathOptInterface.ObjectiveValue — Type.

ObjectiveValue(resultidx::Int=1, objectiveindex::Int=1)

The objective value of the resultindexth primal result of the objectiveindexth objective.

Both resultindex and objectiveindex default to 1.

source

MathOptInterface.ObjectiveBound — Type.

ObjectiveBound()

The best known bound on the optimal objective value.

source

MathOptInterface.RelativeGap — Type.

RelativeGap()

The final relative optimality gap, defined as $\frac{|b-f|}{|f|}$, where $b$ is the best bound and $f$ is the best feasible objective value.

source

MathOptInterface.SolveTime — Type.

SolveTime()

The total elapsed solution time (in seconds) as reported by the solver.

source

MathOptInterface.SimplexIterations — Type.

SimplexIterations()

The cumulative number of simplex iterations during the optimization process. In particular, for a mixed-integer program (MIP), the total simplex iterations for all nodes.

source

MathOptInterface.BarrierIterations — Type.

BarrierIterations()

The cumulative number of barrier iterations while solving a problem.

source

MathOptInterface.NodeCount — Type.

NodeCount()

The total number of branch-and-bound nodes explored while solving a mixed-integer program (MIP).

source

MathOptInterface.TerminationStatus — Type.

TerminationStatus()

A TerminationStatusCode explaining why the solver stopped.

source

MathOptInterface.PrimalStatus — Type.

PrimalStatus(N)
PrimalStatus()

The ResultStatusCode of the primal result N. If N is omitted, it defaults to 1.

source

MathOptInterface.DualStatus — Type.

DualStatus(N)
DualStatus()

The ResultStatusCode of the dual result N. If N is omitted, it defaults to 1.

source

Termination Status

The TerminationStatus attribute indicates why the solver stopped executing. The value of the attribute is of type TerminationStatusCode.

MathOptInterface.TerminationStatusCode — Type.

TerminationStatusCode

An Enum of possible values for the TerminationStatus attribute. This attribute is meant to explain the reason why the solver stopped executing.

OK

These are generally OK statuses.

Success: the algorithm ran successfully and has a result; this includes cases where the algorithm converges to an infeasible point (NLP) or converges to a solution of a homogeneous self-dual problem and has a certificate of primal/dual infeasibility

AlmostSuccess: the algorithm almost ran successfully (e.g., to relaxed convergence tolerances) and has a result

InfeasibleNoResult: the algorithm stopped because it decided that the problem is infeasible but does not have a result to return

UnboundedNoResult: the algorithm stopped because it decided that the problem is unbounded but does not have a result to return

InfeasibleOrUnbounded: the algorithm stopped because it decided that the problem is infeasible or unbounded (no result is available); this occasionally happens during MIP presolve

Limits

The solver stopped because of some user-defined limit. To be documented: IterationLimit, TimeLimit, NodeLimit, SolutionLimit, MemoryLimit, ObjectiveLimit, NormLimit, OtherLimit.

Problematic

This group of statuses means that something unexpected or problematic happened.

SlowProgress: the algorithm stopped because it was unable to continue making progress towards the solution

AlmostSuccess should be used if there is additional information that relaxed convergence tolerances are satisfied

To be documented: NumericalError, InvalidSolverInstance, InvalidOption, Interrupted, OtherError.

source

Result Status

The PrimalStatus and DualStatus attributes indicate how to interpret the result returned by the solver. The value of the attribute is of type ResultStatusCode.

MathOptInterface.ResultStatusCode — Type.

ResultStatusCode

An Enum of possible values for the PrimalStatus and DualStatus attributes. The values indicate how to interpret the result vector.

FeasiblePoint

NearlyFeasiblePoint

InfeasiblePoint

InfeasibilityCertificate

NearlyInfeasibilityCertificate

ReductionCertificate

NearlyReductionCertificate

Unknown

Other

source

Basis Status

[what? BasisStatus]

MathOptInterface.BasisStatusCode — Type.

BasisStatusCode

An Enum of possible values for the VariableBasisStatus and ConstraintBasisStatus attribute. This explains the status of a given element with respect to an optimal solution basis. Possible values are:

Basic: element is in the basis

Nonbasic: element is not in the basis

NonbasicAtLower: element is not in the basis and is at its lower bound

NonbasicAtUpper: element is not in the basis and is at its upper bound

SuperBasic: element is not in the basis but is also not at one of its bounds

source

Sets

List of recognized sets.

MathOptInterface.AbstractSet — Type.

AbstractSet

Abstract supertype for set objects used to encode constraints.

source

MathOptInterface.Reals — Type.

Reals(dim)

The set $\mathbb{R}^{dim}$ (containing all points) of dimension dim.

source

MathOptInterface.Zeros — Type.

Zeros(dim)

The set $\{ 0 \}^{dim}$ (containing only the origin) of dimension dim.

source

MathOptInterface.NonNegatives — Type.

NonNegatives(dim)

The nonnegative orthant $\{ x \in \mathbb{R}^{dim} : x \ge 0 \}$ of dimension dim.

source

MathOptInterface.NonPositives — Type.

NonPositives(dim)

The nonpositive orthant $\{ x \in \mathbb{R}^{dim} : x \le 0 \}$ of dimension dim.

source

MathOptInterface.GreaterThan — Type.

GreaterThan(lower)

The set $[lower,\infty) \subseteq \mathbb{R}$.

source

MathOptInterface.LessThan — Type.

LessThan(upper)

The set $(-\infty,upper] \subseteq \mathbb{R}$.

source

MathOptInterface.Interval — Type.

Interval(lower,upper)

The interval $[lower, upper] \subseteq \mathbb{R}$. If lower or upper is -Inf or Inf, respectively, the set is interpreted as a one-sided interval.

source

MathOptInterface.SecondOrderCone — Type.

SecondOrderCone(dim)

The second-order cone (or Lorenz cone) $\{ (t,x) \in \mathbb{R}^{dim} : t \ge || x ||_2 \}$ of dimension dim.

source

MathOptInterface.ExponentialCone — Type.

ExponentialCone()

The 3-dimensional exponential cone $\{ (x,y,z) \in \mathbb{R}^3 : y \exp (x/y) \le z, y > 0 \}$.

source

MathOptInterface.DualExponentialCone — Type.

DualExponentialCone()

The 3-dimensional dual exponential cone $\{ (u,v,w) \in \mathbb{R}^3 : -u \exp (v/u) \le \exp(1) w, u < 0 \}$.

source

MathOptInterface.PowerCone — Type.

PowerCone(a)

The 3-dimensional power cone $\{ (x,y,z) \in \mathbb{R}^3 : x^{a} y^{1-a} >= |z|, x \ge 0, y \ge 0 \}$ with parameter a.

source

MathOptInterface.DualPowerCone — Type.

DualPowerCone(a)

The 3-dimensional power cone $\{ (u,v,w) \in \mathbb{R}^3 : (\frac{u}{a})^a (\frac{v}/{1-a})^{1-a} >= |w|, u \ge 0, v \ge 0 \}$ with parameter a.

source

MathOptInterface.PositiveSemidefiniteConeTriangle — Type.

PositiveSemidefiniteConeTriangle(dim)

The (vectorized) cone of symmetric positive semidefinite matrices, with off-diagonals unscaled. The entries of the upper triangular part of the matrix are given row by row (or equivalently, the entries of the lower triangular part are given column by column). An $n \times n$ matrix has $n(n+1)/2$ lower-triangular elements, so for the vectorized cone of dimension dim, the corresponding symmetric matrix has side dimension $\sqrt (1/4 + 2 dim) - 1/2$ elements. The scalar product is the sum of the pairwise product of the diagonal entries plus twice the sum of the pairwise product of the upper diagonal entries.

Examples

The matrix

\[\begin{bmatrix} 1 & 2 & 3\\ 2 & 4 & 5\\ 3 & 5 & 6 \end{bmatrix}\]

corresponds to $(1, 2, 3, 4, 5, 6)$ for PositiveSemidefiniteConeTriangle

source

MathOptInterface.PositiveSemidefiniteConeScaled — Type.

PositiveSemidefiniteConeScaled(dim)

The (vectorized) cone of symmetric positive semidefinite matrices, with off-diagonals scaled. The entries of the upper triangular part of the matrix are given row by row (or equivalently, the entries of the lower triangular part are given column by column). An $n \times n$ matrix has $n(n+1)/2$ lower-triangular elements, so for the vectorized cone of dimension dim, the corresponding symmetric matrix has side dimension $\sqrt (1/4 + 2 dim) - 1/2$ elements. The off-diagonal entries of the matrices of both the cone and its dual are scaled by $\sqrt{2}$ and the scalar product is simply the sum of the pairwise product of the entries.

Examples

The matrix

\[\begin{bmatrix} 1 & 2 & 3\\ 2 & 4 & 5\\ 3 & 5 & 6 \end{bmatrix}\]

and to $(1, 2\sqrt{2}, 3\sqrt{2}, 4, 5\sqrt{2}, 6)$ for PositiveSemidefiniteConeScaled.

source

MathOptInterface.Integers — Type.

Integers()

The set of integers $\mathbb{Z}$.

source

MathOptInterface.ZeroOne — Type.

ZeroOne()

The set $\{ 0, 1 \}$.

source

MathOptInterface.SOS1 — Type.

SOS1(weights)

The set corresponding to the special ordered set (SOS) constraint of type 1. Of the variables in the set, at most one can be nonzero. The weights induce an ordering of the variables; as such, they should be unique values. The kth element in the set corresponds to the kth weight in weights. See here for a description of SOS constraints and their potential uses.

source

MathOptInterface.SOS2 — Type.

SOS2(weights)

The set corresponding to the special ordered set (SOS) constraint of type 2. Of the variables in the set, at most two can be nonzero, and if two are nonzero, they must be adjacent in the ordering of the set. The weights induce an ordering of the variables; as such, they should be unique values. The kth element in the set corresponds to the kth weight in weights. See here for a description of SOS constraints and their potential uses.

source

Functions for getting and setting properties of sets.

MathOptInterface.dimension — Function.

dimension(s::AbstractSet)

Return the dimension (number of vector components) in the set s.

source

Variables

Variable references and functions for adding and deleting variables.

[attribute that points to the (scalar) variable domain??? eg GreaterThan, NonNegatives, ZeroOne, SemiInteger]

MathOptInterface.VariableReference — Type.

VariableReference

A lightweight object used to reference variables in a solver instance.

source

MathOptInterface.candelete — Method.

candelete(m::AbstractSolverInstance, ref::VariableReference)::Bool

Return a Bool indicating whether this variable can be removed from the solver instance m.

source

MathOptInterface.isvalid — Method.

isvalid(m::AbstractSolverInstance, ref::VariableReference)::Bool

Return a Bool indicating whether this reference is valid for an active variable in the solver instance m.

source

Base.delete! — Method.

delete!(m::AbstractSolverInstance, ref::VariableReference)

Delete the referenced variable from the solver instance.

delete!(m::AbstractSolverInstance, refs::Vector{VariableReference})

Delete the referenced variables in the vector refs from the solver instance.

source

MathOptInterface.addvariables! — Function.

addvariables!(m::AbstractSolverInstance, n::Int)::Vector{VariableReference}

Add n scalar variables to the solver instance, returning a vector of variable references.

source

MathOptInterface.addvariable! — Function.

addvariable!(m::AbstractSolverInstance)::VariableReference

Add a scalar variable to the solver instance, returning a variable reference. In addition, there is a special case for adding variables to existing linear problems.

addvariable!(m::AbstractSolverInstance, cref::Vector{Union{AffineConstraintRef{NonPositive}, AffineConstraintRef{NonNegative}, AffineConstraintRef{Zero}, AffineConstraintRef{Interval}}}, coefs)::VariableReference

Add a variable with coefficients specified by coefs in the existing affine constraints given by the constraint references cref. To add a variable with coefficients in a constraint that is not listed here, use addvariable!(m) and then modifyconstraint! instead.

source

List of attributes associated with variables. [category AbstractVariableAttribute] Calls to getattribute and setattribute! should include as an argument a single VariableReference or a vector of VariableReference objects.

MathOptInterface.VariablePrimalStart — Type.

VariablePrimalStart()

An initial assignment of the variables that the solver may use to warm-start the solve.

source

MathOptInterface.VariablePrimal — Type.

VariablePrimal(N)
VariablePrimal()

The assignment to the primal variables in result N. If N is omitted, it is 1 by default.

source

MathOptInterface.VariableBasisStatus — Type.

VariableBasisStatus()

Returns the BasisStatusCode of a given variable, with respect to an available optimal solution basis.

source

Objectives

Functions for adding and modifying objectives.

MathOptInterface.setobjective! — Function.

setobjective!(m::AbstractSolverInstance, b, a_varref::Vector{VariableReference}, a_coef, Q_vari::Vector{VariableReference}, Q_varj::Vector{VariableReference}, Q_coef, N::Int=1)

Set the Nth objective in the solver instance m to be $\frac{1}{2} x^T Q_0 x + a_0^T x + b_0$, where:

$a_0$ is a sparse vector specified in tuple form by a_varref, a_coef

$b_0$ is a scalar

the symmetric matrix $Q_0$ is defined by the triplets in Q_vari, Q_varj, Q_coef

Duplicate indices (sparse) in either the $a_0$ vector or the $Q_0$ matrix are accepted and will be summed together. Off-diagonal entries of $Q_0$ will be mirrored, so either the upper triangular or lower triangular entries of $Q_0$ should be provided. If entries for both $(i,j)$ and $(j,i)$ are provided, these are considered duplicate terms. a_varref, Q_vari, Q_varj should be collections of VariableReference objects.

setobjective!(m::AbstractSolverInstance, b, a_varref::Vector{VariableReference}, a_coef, N::Int=1)

Set the Nth objective in the solver instance m to be $a_0^T x + b_0$, where:

$a_0$ is a sparse vector specified in tuple form by a_varref, a_coef

$b_0$ is a scalar

Duplicate indices (sparse) in either the $a_0$ vector or the $Q_0$ matrix are accepted and will be summed together.

source

MathOptInterface.modifyobjective! — Function.

modifyobjective!(m::AbstractSolverInstance, N::Int, args...)

Modify elements of the Nth objective depending on the arguments args. The Nth objective has the form $\frac{1}{2} x^T Q_0 x + a_0^T x + b_0$.

There are three cases, below.

Modify Constant term

modifyobjective!(m::AbstractSolverInstance, N::Int, b)

Set the constant term $b_0$ of the Nth objective to b.

Examples

modifyobjective!(m, 1, 1.0)

Modify Linear term

modifyobjective!(m::AbstractSolverInstance, N::Int, a_varidx, a_coef)

Set elements given by a_varidx in the linear term of the Nth objective to a_coef. Either a_varidx and a_coef are both singletons, or they should be collections with equal length. The behavior of duplicate entries in a_varidx is undefined.

Examples

modifyobjective!(m, 1, v, 1.0)
modifyobjective!(m, 1, [v1, v2], [1.0, 2.0])

Modify Quadratic term

modifyobjective!(m::AbstractSolverInstance, N::Int, Q_vari, Q_varj, Q_coef)

Set the elements in the quadratic term of the Nth objective specified by the triplets Q_vari, Q_varj, Q_coef. Off-diagonal entries will be mirrored. Q_vari, Q_varj should be collections of VariableReference objects. The behavior of duplicate entries is undefined. If entries for both $(i,j)$ and $(j,i)$ are provided, these are considered duplicate terms.

Examples

modifyobjective!(m, 1, v1, v2, 1.0)
modifyobjective!(m, 1, [v1, v2], [v1, v1], [1.0, 2.0])

source

MathOptInterface.getobjectiveconstant — Function.

getobjectiveconstant(m, N::Int=1)

Return the constant term in the Nth objective.

source

MathOptInterface.getobjectiveaffine — Function.

getobjectiveaffine(m, N::Int=1)

Return the affine part of the Nth objective in tuple form (varref, coef) where varref is a VariableReference and coef is a coefficient. Output is a tuple of two vectors.

getobjectiveaffine(m, v::VariableReference, N::Int=1)

Return the coefficient for the variable v in the affine part of the Nth objective.

source

Constraints

Constraint references and functions for adding, modifying, and removing constraints.

MathOptInterface.VariablewiseConstraintReference — Type.

VariablewiseConstraintReference{T}

A lightweight object used to reference variablewise constraints in a solver instance. The parameter T is the type of set constraint referenced.

source

MathOptInterface.AffineConstraintReference — Type.

AffineConstraintReference{T}

A lightweight object used to reference affine-in-set constraints in a solver instance. The parameter T is the type of set constraint referenced.

source

MathOptInterface.QuadraticConstraintReference — Type.

QuadraticConstraintReference{T}

A lightweight object used to reference quadratic-in-set constraints in a solver instance. The parameter T is the type of set constraint referenced.

source

MathOptInterface.candelete — Method.

candelete(m::AbstractSolverInstance, ref::ConstraintReference)::Bool

Return a Bool indicating whether this constraint can be removed from the solver instance m.

source

MathOptInterface.isvalid — Method.

isvalid(m::AbstractSolverInstance, ref::ConstraintReference)::Bool

Return a Bool indicating whether this reference is valid for an active constraint in the solver instance m.

source

Base.delete! — Method.

delete!(m::AbstractSolverInstance, ref::ConstraintReference)

Delete the referenced constraint from the solver instance.

delete!(m::AbstractSolverInstance, refs::Vector{ConstraintReference})

Delete the referenced constraints in the vector refs from the solver instance.

source

MathOptInterface.addconstraint! — Function.

addconstraint!(m::AbstractSolverInstance, b, a_constridx, a_v::Vector{VariableReference}, a_coef, Q_constridx, Q_vari::Vector{VariableReference}, Q_varj::Vector{VariableReference}, Q_coef, S::AbstractSet)::QuadraticConstraintReference{typeof(S)}

Add the vector quadratic-in-set constraint $q_i(x) + A_i^T x + b_i \in \mathcal{S}_i$, where:

$A_i$ is a sparse matrix specified in triplet form by a_constridx, a_v, a_coef

$b_i$ is a vector specified by b

$q_i(x)$ is a vector with component $(q_i(x))_k$ defined as $\frac{1}{2} x^T Q_{i,k} x$, where each symmetric matrix $Q_{i,k}$ has Q_constridx equal to k and is specified in triplet form by Q_vari, Q_varj, Q_coef

$\mathcal{S}_i$ is a pre-defined set specified as S

Duplicate indices in either the $A_i$ matrix or any $Q_{i,k}$ matrix are accepted and will be summed together. Off-diagonal entries of $Q_{i,k}$ will be mirrored, so either the upper triangular or lower triangular entries of $Q_{i,k}$ should be provided. If entries for both $(i,j)$ and $(j,i)$ are provided, these are considered duplicate terms. a_v, Q_vari, Q_varj should be collections of VariableReference objects.

addconstraint!(m::AbstractSolverInstance, b, a_v::Vector{VariableReference}, a_coef, Q_vari::Vector{VariableReference}, Q_varj::Vector{VariableReference}, Q_coef, S::AbstractSet)::QuadraticConstraintReference{typeof(S)}

Add the scalar quadratic-in-set constraint $q_i(x) + a_i^T x + b_i \in \mathcal{S}_i$, where:

$a_i$ is a sparse vector specified in tuple form by a_v, a_coef

$b_i$ is a scalar specified by b

$q_i(x)$ is defined as $\frac{1}{2} x^T Q_{i} x$, where the symmetric matrix $Q_{i}$ is specified in triplet form by Q_vari, Q_varj, Q_coef

$\mathcal{S}_i$ is a pre-defined scalar set specified as S

Duplicate indices in the $a_i$ or the $Q_{i,k}$ are accepted and will be summed together.

addconstraint!(m::AbstractSolverInstance, b, a_constridx, a_v::Vector{VariableReference}, a_coef, S::AbstractSet)::AffineConstraintReference{typeof(S)}

Add the vector affine-in-set constraint $A_i^T x + b_i \in \mathcal{S}_i$, where:

$A_i$ is a sparse matrix specified in triplet form by a_constridx, a_v, a_coef

$b_i$ is a vector specified by b

$\mathcal{S}_i$ is a pre-defined set specified as S

Duplicate indices in the $A_i$ are accepted and will be summed together.

addconstraint!(m::AbstractSolverInstance, b, a_v::Vector{VariableReference}, a_coef, S::AbstractSet)::AffineConstraintReference{typeof(S)}

Add the scalar affine-in-set constraint $a_i^T x + b_i \in \mathcal{S}_i$, where:

$a_i$ is a sparse vector specified in tuple form by a_v, a_coef

$b_i$ is a scalar specified by b

$\mathcal{S}_i$ is a pre-defined scalar set specified as S

Duplicate indices in the $a_i$ are accepted and will be summed together.

addconstraint!(m::AbstractSolverInstance, vs::Vector{VariableReference}, S::AbstractSet)::VariablewiseConstraintReference{typeof(S)}

Add the vector variable-wise constraint $(x_j)_{j \in v_i} \in \mathcal{S}_i$, where:

$v_i$ is a list of variable indices specified as a vector of variable references vs

$\mathcal{S}_i$ is a pre-defined set specified as S

Behavior is not defined for duplicate indices in the $v_i$.

addconstraint!(m::AbstractSolverInstance, v::VariableReference, S::AbstractSet)::VariablewiseConstraintReference{typeof(S)}

Add the scalar variable-wise constraint $x_j \in \mathcal{S}_i$, where:

$x_j$ is variable specified as a variable reference v

$\mathcal{S}_i$ is a pre-defined scalar set specified as S

source

MathOptInterface.modifyconstraint! — Function.

modifyconstraint!(m::AbstractSolverInstance, c::ConstraintReference, k::Int, args...)

Modify elements of the kth row of the constraint c depending on the arguments args. The kth row will have the form $q_{i,k}(x) + A_{i,k}^T x + b_{i,k}$. There are four cases.

Modify Constant term

modifyconstraint!(m::AbstractSolverInstance, c::ConstraintReference, k::Int, b)

Set the constant term of the kth row in the constraint c to b.

Examples

modifyconstraint!(m, c, 1, 1.0)

Modify Linear term

modifyconstraint!(m::AbstractSolverInstance, c::ConstraintReference, k::Int, a_v::Vector{VariableReference}, a_coef)

Set elements given by a_v in the linear term of the kth row in the constraint c to a_coef. Either a_v and a_coef are both singletons, or they should be collections with equal length. The behavior of duplicate entries in a_v is undefined.

Examples

modifyconstraint!(m, c, v, 1.0)
modifyconstraint!(m, c, [v_1, v_2], [1.0, 2.0])

Modify Quadratic term

modifyconstraint!(m::AbstractSolverInstance, c::ConstraintReference, k::Int, Q_vari, Q_varj, Q_coef)

Set the elements in the quadratic term of the kth row of the constraint c specified by the triplets Q_vari, Q_varj, Q_coef. Off-diagonal entries will be mirrored. Q_vari, Q_varj should be collections of VariableReference objects. The behavior of duplicate entries is undefined. If entries for both $(i,j)$ and $(j,i)$ are provided, these are considered duplicate terms.

Examples

modifyconstraint!(m, c, v_1, v_2, 1.0)
modifyconstraint!(m, c, [v_1, v_2], [v_1, v_1], [1.0, 2.0])

Modify Set

modifyconstraint!(m::AbstractSolverInstance, c::ConstraintReference{Set}, S::Set)

Change the set of constraint c to the new set S which should be of the same type as the original set.

Examples

If c is a ConstraintReference{Interval}

modifyconstraint!(m, c, Interval(0, 5))
modifyconstraint!(m, c, NonPositives) # errors

source

MathOptInterface.getconstraintconstant — Function.

getconstraintconstant(m::AbstractSolverInstance, c::ConstraintReference)

Return the $b$ vector of the constraint c.

getconstraintconstant(m::AbstractSolverInstance, c::ConstraintReference, k::Int)

Return the constant term of the kth row of the constraint c.

source

MathOptInterface.getconstraintaffine — Function.

getconstraintaffine(m::AbstractSolverInstance, c::ConstraintReference)

Return the $A_i$ matrix of the constraint corresponding to c in triplet form (i, v, coef), where v is a VariableReference, and coef is a coefficient value. Output is a tuple of three vectors.

getconstraintaffine(m::AbstractSolverInstance, c::ConstraintReference, k::Int)

Return the kth row of the $A_k$ matrix of the constraint corresponding to c in tuple form (v, coef), where v is a VariableReference, and coef is a coefficient value. Output is a tuple of two vectors.

getconstraintaffine(m::AbstractSolverInstance, c::ConstraintReference, k::Int, v::VariableReference)

Return the element of the $A_k$ matrix of the constraint corresponding to c in row k and variable v.

source

MathOptInterface.getconstraintquadratic — Function.

getconstraintquadratic(m::AbstractSolverInstance, c::ConstraintReference, k::Int)

Return the $Q_{i,k}$ matrix of the kth row of the constraint corresponding to c in triplet form (v_1, v_2, coef), where v_1, v_2 are VariableReferences, and coef is a coefficient value. Output is a tuple of three vectors. The $Q_{i,k}$ matrix must be symmetric, and only one element is returned.

getconstraintquadratic(m::AbstractSolverInstance, c::ConstraintReference, k::Int, v_1::VariableReference, v_2::VariableReference)

Return the element corresponding to (v_1, v_2) of the $Q_{i,k}$ matrix of the kth row of the constraint corresponding to c.

source

List of attributes associated with constraints. [category AbstractConstraintAttribute] Calls to getattribute and setattribute! should include as an argument a single ConstraintReference or a vector of ConstraintReference{T} objects.

[why is ConstraintBasisStatus under constraint attributes but below we have a basis status attribute separately??]

MathOptInterface.ConstraintPrimalStart — Type.

ConstraintPrimalStart()

An initial assignment of the constraint primal values that the solver may use to warm-start the solve.

source

MathOptInterface.ConstraintDualStart — Type.

ConstraintDualStart()

An initial assignment of the constriant duals that the solver may use to warm-start the solve.

source

MathOptInterface.ConstraintPrimal — Type.

ConstraintPrimal(N)
ConstraintPrimal()

The assignment to the constraint primal values in result N. If N is omitted, it is 1 by default.

source

MathOptInterface.ConstraintDual — Type.

ConstraintDual(N)
ConstraintDual()

The assignment to the constraint dual values in result N. If N is omitted, it is 1 by default.

source

MathOptInterface.ConstraintBasisStatus — Type.

ConstraintBasisStatus()

Returns the BasisStatusCode of a given constraint, with respect to an available optimal solution basis.

source