Reference

API Reference

[Some introduction to API. List basic standalone methods.]

Parameters and Attributes

Parameter abstract type.

MathOptInterface.AbstractSolverParameter — Type.

AbstractSolverParameter

Abstract supertype for parameter objects that can be used to set or get parameters of the solver.

Note

The difference between AbstractSolverParameter and AbstractInstanceAttribute lies in the behavior of isempty, empty! and copy!. Typically solver parameters only affect how the instance is solved.

source

List of attribute categories.

MathOptInterface.AbstractInstanceAttribute — Type.

AbstractInstanceAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of the instance.

source

MathOptInterface.AbstractVariableAttribute — Type.

AbstractVariableAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of variables in the instance.

source

MathOptInterface.AbstractConstraintAttribute — Type.

AbstractConstraintAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of constraints in the instance.

source

Functions for getting and setting parameters and attributes.

MathOptInterface.canget — Function.

canget(instance::AbstractInstance, param::AbstractSolverParameter)::Bool

Return a Bool indicating whether instance currently has a value for the parameter specified by parameter type param.

canget(instance::AbstractInstance, attr::AbstractInstanceAttribute)::Bool

Return a Bool indicating whether instance currently has a value for the attribute specified by attribute type attr.

canget(instance::AbstractInstance, attr::AbstractVariableAttribute, ::Type{VariableIndex})::Bool

Return a Bool indicating whether instance currently has a value for the attribute specified by attribute type attr applied to every variable of the instance.

canget(instance::AbstractInstance, attr::AbstractConstraintAttribute, ::Type{ConstraintIndex{F,S}})::Bool where {F<:AbstractFunction,S<:AbstractSet}

Return a Bool indicating whether instance currently has a value for the attribute specified by attribute type attr applied to every F-in-S constraint.

canget(instance::AbstractInstance, ::Type{VariableIndex}, name::String)::Bool

Return a Bool indicating if a variable with the name name exists in instance.

canget(instance::AbstractInstance, ::Type{ConstraintIndex{F,S}}, name::String)::Bool where {F<:AbstractFunction,S<:AbstractSet}

Return a Bool indicating if an F-in-S constraint with the name name exists in instance.

canget(instance::AbstractInstance, ::Type{ConstraintIndex}, name::String)::Bool

Return a Bool indicating if a constraint of any kind with the name name exists in instance.

Examples

canget(instance, ObjectiveValue())
canget(instance, VariablePrimalStart(), VariableIndex)
canget(instance, VariablePrimal(), VariableIndex)
canget(instance, ConstraintPrimal(), ConstraintIndex{SingleVariable,EqualTo{Float64}})
canget(instance, VariableIndex, "var1")
canget(instance, ConstraintIndex{ScalarAffineFunction{Float64},LessThan{Float64}}, "con1")
canget(instance, ConstraintIndex, "con1")

source

MathOptInterface.get — Function.

get(instance::AbstractInstance, param::AbstractSolverParameter)

Return a parameter param of the instance instance.

get(instance::AbstractInstance, attr::AbstractInstanceAttribute)

Return an attribute attr of the instance instance.

get(instance::AbstractInstance, attr::AbstractVariableAttribute, v::VariableIndex)

Return an attribute attr of the variable v in instance instance.

get(instance::AbstractInstance, attr::AbstractVariableAttribute, v::Vector{VariableIndex})

Return a vector of attributes corresponding to each variable in the collection v in the instance instance.

get(instance::AbstractInstance, attr::AbstractConstraintAttribute, c::ConstraintIndex)

Return an attribute attr of the constraint c in instance instance.

get(instance::AbstractInstance, attr::AbstractConstraintAttribute, c::Vector{ConstraintIndex{F,S}})

Return a vector of attributes corresponding to each constraint in the collection c in the instance instance.

get(instance::AbstractInstance, ::Type{VariableIndex}, name::String)

If a variable with name name exists in the instance instance, return the corresponding index, otherwise throw a KeyError.

get(instance::AbstractInstance, ::Type{ConstraintIndex{F,S}}, name::String) where {F<:AbstractFunction,S<:AbstractSet}

If an F-in-S constraint with name name exists in the instance instance, return the corresponding index, otherwise throw a KeyError.

get(instance::AbstractInstance, ::Type{ConstraintIndex}, name::String)

If any constraint with name name exists in the instance instance, return the corresponding index, otherwise throw a KeyError. This version is available for convenience but may incur a performance penalty because it is not type stable.

Examples

get(instance, ObjectiveValue())
get(instance, VariablePrimal(), ref)
get(instance, VariablePrimal(5), [ref1, ref2])
get(instance, OtherAttribute("something specific to cplex"))
get(instance, VariableIndex, "var1")
get(instance, ConstraintIndex{ScalarAffineFunction{Float64},LessThan{Float64}}, "con1")
get(instance, ConstraintIndex, "con1")

source

MathOptInterface.get! — Function.

get!(output, instance::AbstractInstance, args...)

An in-place version of get. The signature matches that of get except that the the result is placed in the vector output.

source

MathOptInterface.canset — Function.

canset(instance::AbstractInstance, param::AbstractSolverParameter)::Bool

Return a Bool indicating whether it is possible to set the parameter param to the instance instance.

canset(instance::AbstractInstance, attr::AbstractInstanceAttribute)::Bool

Return a Bool indicating whether it is possible to set the attribute attr to the instance instance.

canset(instance::AbstractInstance, attr::AbstractVariableAttribute, R::Type{VariableIndex})::Bool
canset(instance::AbstractInstance, attr::AbstractConstraintAttribute, R::Type{ConstraintIndex{F,S})::Bool

Return a Bool indicating whether it is possible to set attribute attr applied to the index type R in the instance instance.

Examples

canset(instance, ObjectiveValue())
canset(instance, VariablePrimalStart(), VariableIndex)
canset(instance, ConstraintPrimal(), ConstraintIndex{VectorAffineFunction{Float64},Nonnegatives})

source

MathOptInterface.set! — Function.

set!(instance::AbstractInstance, param::AbstractSolverParameter, value)

Assign value to the parameter param of the instance instance.

set!(instance::AbstractInstance, attr::AbstractInstanceAttribute, value)

Assign value to the attribute attr of the instance instance.

set!(instance::AbstractInstance, attr::AbstractVariableAttribute, v::VariableIndex, value)

Assign value to the attribute attr of variable v in instance instance.

set!(instance::AbstractInstance, attr::AbstractVariableAttribute, v::Vector{VariableIndex}, vector_of_values)

Assign a value respectively to the attribute attr of each variable in the collection v in instance instance.

set!(instance::AbstractInstance, attr::AbstractConstraintAttribute, c::ConstraintIndex, value)

Assign a value to the attribute attr of constraint c in instance instance.

set!(instance::AbstractInstance, attr::AbstractConstraintAttribute, c::Vector{ConstraintIndex{F,S}}, vector_of_values)

Assign a value respectively to the attribute attr of each constraint in the collection c in instance instance.

source

MathOptInterface.supports — Function.

supports(instance::AbstractInstance, param::AbstractSolverParameter)::Bool

Return a Bool indicating whether instance supports the solver parameter param.

supports(instance::AbstractInstance, attr::AbstractInstanceAttribute)::Bool

Return a Bool indicating whether instance supports the instance attribute attr.

supports(instance::AbstractInstance, attr::AbstractVariableAttribute, ::Type{VariableIndex})::Bool

Return a Bool indicating whether instance supports the variable attribute attr.

supports(instance::AbstractInstance, attr::AbstractConstraintAttribute, ::Type{ConstraintIndex{F,S}})::Bool where {F,S}

Return a Bool indicating whether instance supports the constraint attribute attr applied to an F-in-S constraint.

In other words, it should return true if copy!(instance, src) does not return CopyUnsupportedAttribute when the attribute attr is set to src. If the attribute is only not supported in specific circumstances, it should still return true.

source

Instance

MathOptInterface.AbstractInstance — Type.

AbstractInstance

Abstract supertype for objects representing an instance of an optimization problem.

source

MathOptInterface.AbstractStandaloneInstance — Type.

AbstractStandaloneInstance

Abstract supertype for objects representing an instance of an optimization problem unattached to any particular solver. Does not have methods for solving or querying results.

source

MathOptInterface.AbstractSolverInstance — Type.

AbstractSolverInstance

Abstract supertype for objects representing an instance of an optimization problem tied to a particular solver. This is typically a solver's in-memory representation. In addition to AbstractInstance, AbstractSolverInstance objects let you solve the instance and query the solution.

source

MathOptInterface.isempty — Function.

isempty(instance::AbstractInstance)

Returns false if the instance has any instance attribute set or has any variables or constraints. Note that an empty instance can have solver parameters set.

source

MathOptInterface.empty! — Function.

empty!(instance::AbstractInstance)

Empty the instance, that is, remove from the instance instance all variables, constraints and instance attributes but not solver parameters.

source

MathOptInterface.write — Function.

write(instance::AbstractInstance, filename::String)

Writes the current instance data to the given file. Supported file types depend on the solver or standalone instance type.

source

MathOptInterface.read! — Function.

read!(instance::AbstractInstance, filename::String)

Read the file filename into the instance instance. If m is non-empty, this may throw an error.

Supported file types depend on the instance type.

source

Copying

MathOptInterface.copy! — Function.

copy!(dest::AbstractInstance, src::AbstractInstance, warnattributes=true)::CopyResult

Copy the model from the instance src into the instance dest. The target instance dest is emptied, and all previous indices to variables or constraints in dest are invalidated. Returns a CopyResult object. If the copy is successfully, the CopyResult contains a dictionary-like object that translates variable and constraint indices from the src instance to the corresponding indices in the dest instance.

If an attribute attr cannot be copied from src to dest then an error is thrown. If a solver parameter cannot be copied then:

If warnattributes is true, a warning is displayed, otherwise,

The attribute is silently ignored.

Example

# Given empty `AbstractInstance`s `src` and `dest`.

x = addvariable!(src)

isvalid(src, x)   # true
isvalid(dest, x)  # false (`dest` has no variables)

copy_result = copy!(dest, src)
if copy_result.status == CopySuccess
    index_map = copy_result.indexmap
    isvalid(dest, x) # false (unless index_map[x] == x)
    isvalid(dest, index_map[x]) # true
else
    println("Copy failed with status ", copy_result.status)
    println("Failure message: ", copy_result.message)
end

source

MathOptInterface.CopyResult — Type.

struct CopyResult{T}
    status::CopyStatusCode
    message::String # Human-friendly explanation why the copy failed
    indexmap::T     # Only valid if status is CopySuccess
end

A struct returned by copy! to indicate success or failure. If success, also exposes a map between the variable and constraint indices of the two instances.

source

MathOptInterface.CopyStatusCode — Type.

CopyStatusCode

An Enum of possible statuses returned by a copy! operation through the CopyResult struct.

CopySuccess: The copy was successful.

CopyUnsupportedAttribute: The copy failed because the destination does not support an attribute present in the source.

CopyUnsupportedConstraint: The copy failed because the destination does not support a constraint present in the source.

CopyOtherError: The copy failed for a different reason.

In the failure cases:

See the corresponding message field of the CopyResult for an explanation of the failure.

The state of the destination instance is undefined.

source

List of instance attributes

MathOptInterface.Name — Type.

Name()

A string identifying the instance.

source

MathOptInterface.ObjectiveSense — Type.

ObjectiveSense()

The sense of the objective function, an OptimizationSense with value MinSense, MaxSense, or FeasiblitySense.

source

MathOptInterface.NumberOfVariables — Type.

NumberOfVariables()

The number of variables in the instance.

source

MathOptInterface.ListOfVariableIndices — Type.

ListOfVariableIndices()

A Vector{VariableIndex} indexing all variables present in the instance (i.e., of length equal to the value of NumberOfVariables()).

source

MathOptInterface.ListOfConstraints — Type.

ListOfConstraints()

A list of tuples of the form (F,S), where F is a function type and S is a set type indicating that the attribute NumberOfConstraints{F,S}() has value greater than zero.

source

MathOptInterface.NumberOfConstraints — Type.

NumberOfConstraints{F,S}()

The number of constraints of the type F-in-S present in the instance.

source

MathOptInterface.ListOfConstraintIndices — Type.

ListOfConstraintIndices{F,S}()

A Vector{ConstraintIndex{F,S}} indexing all constraints of type F-inS in the instance (i.e., of length equal to the value of NumberOfConstraints{F,S}()).

source

MathOptInterface.ListOfInstanceAttributesSet — Type.

ListOfInstanceAttributesSet()

A Vector{AbstractInstanceAttribute} of all instance attributes that were set to the instance.

source

MathOptInterface.ListOfVariableAttributesSet — Type.

ListOfVariableAttributesSet()

A Vector{AbstractVariableAttribute} of all variable attributes that were set to the instance.

source

MathOptInterface.ListOfConstraintAttributesSet — Type.

ListOfConstraintAttributesSet{F, S}()

A Vector{AbstractConstraintAttribute} of all constraint attributes that were set to F-in-S constraints.

source

There are no attributes specific to a standalone instance.

Solver instance

MathOptInterface.optimize! — Function.

optimize!(instance::AbstractSolverInstance)

Start the solution procedure.

source

MathOptInterface.free! — Function.

free!(instance::AbstractSolverInstance)

Release any resources and memory used by the solver instance. Note that the Julia garbage collector takes care of this automatically, but automatic collection cannot always be forced. This method is useful for more precise control of resources, especially in the case of commercial solvers with licensing restrictions on the number of concurrent runs. Users must discard the solver instance object after this method is invoked.

source

List of solver instance attributes

MathOptInterface.RawSolver — Type.

RawSolver()

An object that may be used to access a solver-specific API for this solver instance.

source

MathOptInterface.ResultCount — Type.

ResultCount()

The number of results available.

source

MathOptInterface.ObjectiveFunction — Type.

ObjectiveFunction{F<:AbstractScalarFunction}()

An F instance which represents the objective function. It is guaranteed to be equivalent but not necessarily identical to the function provided by the user. Throws an InexactError if the objective function cannot be converted to F, e.g. the objective function is quadratic and F is ScalarAffineFunction{Float64} or it has non-integer coefficient and F is ScalarAffineFunction{Int}.

source

MathOptInterface.ObjectiveValue — Type.

ObjectiveValue(resultidx::Int=1)

The objective value of the resultindexth primal result.

source

MathOptInterface.ObjectiveBound — Type.

ObjectiveBound()

The best known bound on the optimal objective value.

source

MathOptInterface.RelativeGap — Type.

RelativeGap()

The final relative optimality gap, defined as $\frac{|b-f|}{|f|}$, where $b$ is the best bound and $f$ is the best feasible objective value.

source

MathOptInterface.SolveTime — Type.

SolveTime()

The total elapsed solution time (in seconds) as reported by the solver.

source

MathOptInterface.SimplexIterations — Type.

SimplexIterations()

The cumulative number of simplex iterations during the optimization process. In particular, for a mixed-integer program (MIP), the total simplex iterations for all nodes.

source

MathOptInterface.BarrierIterations — Type.

BarrierIterations()

The cumulative number of barrier iterations while solving a problem.

source

MathOptInterface.NodeCount — Type.

NodeCount()

The total number of branch-and-bound nodes explored while solving a mixed-integer program (MIP).

source

MathOptInterface.TerminationStatus — Type.

TerminationStatus()

A TerminationStatusCode explaining why the solver stopped.

source

MathOptInterface.PrimalStatus — Type.

PrimalStatus(N)
PrimalStatus()

The ResultStatusCode of the primal result N. If N is omitted, it defaults to 1.

source

MathOptInterface.DualStatus — Type.

DualStatus(N)
DualStatus()

The ResultStatusCode of the dual result N. If N is omitted, it defaults to 1.

source

Termination Status

The TerminationStatus attribute indicates why the solver stopped executing. The value of the attribute is of type TerminationStatusCode.

MathOptInterface.TerminationStatusCode — Type.

TerminationStatusCode

An Enum of possible values for the TerminationStatus attribute. This attribute is meant to explain the reason why the solver stopped executing.

OK

These are generally OK statuses.

Success: the algorithm ran successfully and has a result; this includes cases where the algorithm converges to an infeasible point (NLP) or converges to a solution of a homogeneous self-dual problem and has a certificate of primal/dual infeasibility

InfeasibleNoResult: the algorithm stopped because it decided that the problem is infeasible but does not have a result to return

UnboundedNoResult: the algorithm stopped because it decided that the problem is unbounded but does not have a result to return

InfeasibleOrUnbounded: the algorithm stopped because it decided that the problem is infeasible or unbounded (no result is available); this occasionally happens during MIP presolve

Limits

The solver stopped because of some user-defined limit. To be documented: IterationLimit, TimeLimit, NodeLimit, SolutionLimit, MemoryLimit, ObjectiveLimit, NormLimit, OtherLimit.

Problematic

This group of statuses means that something unexpected or problematic happened.

SlowProgress: the algorithm stopped because it was unable to continue making progress towards the solution

AlmostSuccess should be used if there is additional information that relaxed convergence tolerances are satisfied

To be documented: NumericalError, InvalidInstance, InvalidOption, Interrupted, OtherError.

source

Result Status

The PrimalStatus and DualStatus attributes indicate how to interpret the result returned by the solver. The value of the attribute is of type ResultStatusCode.

MathOptInterface.ResultStatusCode — Type.

ResultStatusCode

An Enum of possible values for the PrimalStatus and DualStatus attributes. The values indicate how to interpret the result vector.

FeasiblePoint

NearlyFeasiblePoint

InfeasiblePoint

InfeasibilityCertificate

NearlyInfeasibilityCertificate

ReductionCertificate

NearlyReductionCertificate

UnknownResultStatus

OtherResultStatus

source

Variables and Constraints

Basis Status

The BasisStatus attribute of a variable or constraint describes its status with respect to a basis, if one is known. The value of the attribute is of type BasisStatusCode.

MathOptInterface.BasisStatusCode — Type.

BasisStatusCode

An Enum of possible values for the VariableBasisStatus and ConstraintBasisStatus attribute. This explains the status of a given element with respect to an optimal solution basis. Possible values are:

Basic: element is in the basis

Nonbasic: element is not in the basis

NonbasicAtLower: element is not in the basis and is at its lower bound

NonbasicAtUpper: element is not in the basis and is at its upper bound

SuperBasic: element is not in the basis but is also not at one of its bounds

source

Index types

MathOptInterface.VariableIndex — Type.

VariableIndex

A type-safe wrapper for Int64 for use in referencing variables in an instance. To allow for deletion, indices need not be consecutive.

source

MathOptInterface.ConstraintIndex — Type.

ConstraintIndex{F,S}

A type-safe wrapper for Int64 for use in referencing F-in-S constraints in an instance. The parameter F is the type of the function in the constraint, and the parameter S is the type of set in the constraint. To allow for deletion, indices need not be consecutive.

source

MathOptInterface.candelete — Function.

candelete(instance::AbstractInstance, index::Index)::Bool

Return a Bool indicating whether the object referred to by index can be removed from the instance instance.

source

MathOptInterface.isvalid — Function.

isvalid(instance::AbstractInstance, index::Index)::Bool

Return a Bool indicating whether this index refers to a valid object in the instance instance.

source

Base.delete! — Method.

delete!(instance::AbstractInstance, index::Index)

Delete the referenced object from the instance.

delete!{R}(instance::AbstractInstance, indices::Vector{R<:Index})

Delete the referenced objects in the vector indices from the instance. It may be assumed that R is a concrete type.

source

Variables

Functions for adding variables. For deleting, see index types section.

MathOptInterface.canaddvariable — Function.

canaddvariable(instance::AbstractInstance)::Bool

Return a Bool indicating whether it is possible to add a variable to the instance instance.

source

MathOptInterface.addvariables! — Function.

addvariables!(instance::AbstractInstance, n::Int)::Vector{VariableIndex}

Add n scalar variables to the instance, returning a vector of variable indices.

source

MathOptInterface.addvariable! — Function.

addvariable!(instance::AbstractInstance)::VariableIndex

Add a scalar variable to the instance, returning a variable index.

source

List of attributes associated with variables. [category AbstractVariableAttribute] Calls to get and set! should include as an argument a single VariableIndex or a vector of VariableIndex objects.

MathOptInterface.VariableName — Type.

VariableName()

A string identifying the variable. It is invalid for two variables to have the same name.

source

MathOptInterface.VariablePrimalStart — Type.

VariablePrimalStart()

An initial assignment of the variables that the solver may use to warm-start the solve.

source

MathOptInterface.VariablePrimal — Type.

VariablePrimal(N)
VariablePrimal()

The assignment to the primal variables in result N. If N is omitted, it is 1 by default.

source

MathOptInterface.VariableBasisStatus — Type.

VariableBasisStatus()

Returns the BasisStatusCode of a given variable, with respect to an available optimal solution basis.

source

Constraints

Functions for adding and modifying constraints.

MathOptInterface.isvalid — Method.

isvalid(instance::AbstractInstance, index::Index)::Bool

Return a Bool indicating whether this index refers to a valid object in the instance instance.

source

MathOptInterface.canaddconstraint — Function.

canaddconstraint(instance::AbstractInstance, ::Type{F}, ::Type{S})::Bool where {F<:AbstractFunction,S<:AbstractSet}

Return a Bool indicating whether it is possible to add a constraint $f(x) \in \mathcal{S}$ where $f$ is of type F, and $\mathcal{S}$ is of type S.

source

MathOptInterface.addconstraint! — Function.

addconstraint!(instance::AbstractInstance, func::F, set::S)::ConstraintIndex{F,S} where {F,S}

Add the constraint $f(x) \in \mathcal{S}$ where $f$ is defined by func, and $\mathcal{S}$ is defined by set.

addconstraint!(instance::AbstractInstance, v::VariableIndex, set::S)::ConstraintIndex{SingleVariable,S} where {S}
addconstraint!(instance::AbstractInstance, vec::Vector{VariableIndex}, set::S)::ConstraintIndex{VectorOfVariables,S} where {S}

Add the constraint $v \in \mathcal{S}$ where $v$ is the variable (or vector of variables) referenced by v and $\mathcal{S}$ is defined by set.

source

MathOptInterface.addconstraints! — Function.

addconstraints!(instance::AbstractInstance, funcs::Vector{F}, sets::Vector{S})::Vector{ConstraintIndex{F,S}} where {F,S}

Add the set of constraints specified by each function-set pair in funcs and sets. F and S should be concrete types. This call is equivalent to addconstraint!.(instance, funcs, sets) but may be more efficient.

source

MathOptInterface.modifyconstraint! — Function.

Modify Function

modifyconstraint!(instance::AbstractInstance, c::ConstraintIndex{F,S}, func::F)

Replace the function in constraint c with func. F must match the original function type used to define the constraint.

Examples

If c is a ConstraintIndex{ScalarAffineFunction,S} and v1 and v2 are VariableIndex objects,

modifyconstraint!(instance, c, ScalarAffineFunction([v1,v2],[1.0,2.0],5.0))
modifyconstraint!(instance, c, SingleVariable(v1)) # Error

Modify Set

modifyconstraint!(instance::AbstractInstance, c::ConstraintIndex{F,S}, set::S)

Change the set of constraint c to the new set set which should be of the same type as the original set.

Examples

If c is a ConstraintIndex{F,Interval}

modifyconstraint!(instance, c, Interval(0, 5))
modifyconstraint!(instance, c, NonPositives) # Error

Partial Modifications

modifyconstraint!(instance::AbstractInstance, c::ConstraintIndex, change::AbstractFunctionModification)

Apply the modification specified by change to the function of constraint c.

Examples

modifyconstraint!(instance, c, ScalarConstantChange(10.0))

source

MathOptInterface.canmodifyconstraint — Function.

Modify Function

canmodifyconstraint(instance::AbstractInstance, c::ConstraintIndex{F,S}, ::Type{F})::Bool

Return a Bool indicating whether the function in constraint c can be replaced by another function of the same type F as the original function.

Modify Set

canmodifyconstraint(instance::AbstractInstance, c::ConstraintIndex{F,S}, ::Type{S})::Bool

Return a Bool indicating whether the set in constraint c can be replaced by another set of the same type S as the original set.

Partial Modifications

canmodifyconstraint(instance::AbstractInstance, c::ConstraintIndex, ::Type{M})::Bool where M<:AbstractFunctionModification

Return a Bool indicating whether it is possible to apply a modification of type M to the function of constraint c.

Examples

canmodifyconstraint(instance, c, ScalarConstantChange{Float64})

source

MathOptInterface.transformconstraint! — Function.

Transform Constraint Set

transformconstraint!(instance::AbstractInstance, c::ConstraintIndex{F,S1}, newset::S2)::ConstraintIndex{F,S2}

Replace the set in constraint c with newset. The constraint index c will no longer be valid, and the function returns a new constraint index with the correct type.

Solvers may only support a subset of constraint transforms that they perform efficiently (for example, changing from a LessThan to GreaterThan set). In addition, set modification (where S1 = S2) should be performed via the modifyconstraint! function.

Typically, the user should delete the constraint and add a new one.

Examples

If c is a ConstraintIndex{ScalarAffineFunction{Float64},LessThan{Float64}},

c2 = transformconstraint!(instance, c, GreaterThan(0.0))
transformconstraint!(instance, c, LessThan(0.0)) # errors

source

MathOptInterface.cantransformconstraint — Function.

Transform Constraint Set

cantransformconstraint(instance::AbstractInstance, c::ConstraintIndex{F,S1}, ::Type{S2})::Bool where S2<:AbstractSet

Return a Bool indicating whether the set of type S1 in constraint c can be replaced by a set of type S2.

Examples

If c is a ConstraintIndex{ScalarAffineFunction{Float64},LessThan{Float64}},

cantransformconstraint(instance, c, GreaterThan(0.0)) # true
cantransformconstraint(instance, c, ZeroOne())        # false

source

MathOptInterface.supportsconstraint — Function.

supportsconstraint(instance::AbstractInstance, ::Type{F}, ::Type{S})::Bool where {F<:AbstractFunction,S<:AbstractSet}

Return a Bool indicating whether instance supports F-in-S constraints, that is, copy!(instance, src) does not return CopyUnsupportedConstraint when src contains F-in-S constraints. If F-in-S constraints are only not supported in specific circumstances, e.g. F-in-S constraints cannot be combined with another type of constraint, it should still return true.

source

List of attributes associated with constraints. [category AbstractConstraintAttribute] Calls to get and set! should include as an argument a single ConstraintIndex or a vector of ConstraintIndex{F,S} objects.

MathOptInterface.ConstraintName — Type.

ConstraintName()

A string identifying the constraint. It is invalid for two constraints of any kind to have the same name.

source

MathOptInterface.ConstraintPrimalStart — Type.

ConstraintPrimalStart()

An initial assignment of the constraint primal values that the solver may use to warm-start the solve.

source

MathOptInterface.ConstraintDualStart — Type.

ConstraintDualStart()

An initial assignment of the constraint duals that the solver may use to warm-start the solve.

source

MathOptInterface.ConstraintPrimal — Type.

ConstraintPrimal(N)
ConstraintPrimal()

The assignment to the constraint primal values in result N. If N is omitted, it is 1 by default.

source

MathOptInterface.ConstraintDual — Type.

ConstraintDual(N)
ConstraintDual()

The assignment to the constraint dual values in result N. If N is omitted, it is 1 by default.

source

MathOptInterface.ConstraintBasisStatus — Type.

ConstraintBasisStatus()

Returns the BasisStatusCode of a given constraint, with respect to an available optimal solution basis.

source

MathOptInterface.ConstraintFunction — Type.

ConstraintFunction()

Return the AbstractFunction object used to define the constraint. It is guaranteed to be equivalent but not necessarily identical to the function provided by the user.

source

MathOptInterface.ConstraintSet — Type.

ConstraintSet()

Return the AbstractSet object used to define the constraint.

source

Functions and function modifications

List of recognized functions.

MathOptInterface.AbstractFunction — Type.

AbstractFunction

Abstract supertype for function objects.

source

MathOptInterface.SingleVariable — Type.

SingleVariable(variable)

The function that extracts the scalar variable referenced by variable, a VariableIndex. This function is naturally be used for single variable bounds or integrality constraints.

source

MathOptInterface.VectorOfVariables — Type.

VectorOfVariables(variables)

The function that extracts the vector of variables referenced by variables, a Vector{VariableIndex}. This function is naturally be used for constraints that apply to groups of variables, such as an "all different" constraint, an indicator constraint, or a complementarity constraint.

source

MathOptInterface.ScalarAffineFunction — Type.

ScalarAffineFunction{T}(variables, coefficients, constant)

The scalar-valued affine function $a^T x + b$, where:

$a$ is a sparse vector specified in tuple form by variables::Vector{VariableIndex} and coefficients::Vector{T}

$b$ is a scalar specified by constant::T

Duplicate variable indices in variables are accepted, and the corresponding coefficients are summed together.

source

MathOptInterface.VectorAffineFunction — Type.

VectorAffineFunction{T}(outputindex, variables, coefficients, constant)

The vector-valued affine function $A x + b$, where:

$A$ is a sparse matrix specified in triplet form by outputindex, variables, coefficients

$b$ is a vector specified by constant

Duplicate indices in the $A$ are accepted, and the corresponding coefficients are summed together.

source

MathOptInterface.ScalarQuadraticFunction — Type.

ScalarQuadraticFunction{T}(affine_variables, affine_coefficients, quadratic_rowvariables, quadratic_colvariables, quadratic_coefficients, constant)

The scalar-valued quadratic function $\frac{1}{2}x^TQx + a^T x + b$, where:

$a$ is a sparse vector specified in tuple form by affine_variables, affine_coefficients

$b$ is a scalar specified by constant

$Q$ is a symmetric matrix is specified in triplet form by quadratic_rowvariables, quadratic_colvariables, quadratic_coefficients

Duplicate indices in $a$ or $Q$ are accepted, and the corresponding coefficients are summed together. "Mirrored" indices (q,r) and (r,q) (where r and q are VariableIndexes) are considered duplicates; only one need be specified.

source

MathOptInterface.VectorQuadraticFunction — Type.

VectorQuadraticFunction{T}(affine_outputindex, affine_variables, affine_coefficients, quadratic_outputindex, quadratic_rowvariables, quadratic_colvariables, quadratic_coefficients, constant)

The vector-valued quadratic function with ith component ("output index") defined as $\frac{1}{2}x^TQ_ix + a_i^T x + b_i$, where:

$a_i$ is a sparse vector specified in tuple form by the subset of affine_variables, affine_coefficients for the indices k where affine_outputindex[k] == i.

$b_i$ is a scalar specified by constant[i]

$Q_i$ is a symmetric matrix is specified in triplet form by the subset of quadratic_rowvariables, quadratic_colvariables, quadratic_coefficients for the indices k where quadratic_outputindex[k] == i

Duplicate indices in $a_i$ or $Q_i$ are accepted, and the corresponding coefficients are summed together. "Mirrored" indices (q,r) and (r,q) (where r and q are VariableIndexes) are considered duplicates; only one need be specified.

source

List of function modifications.

MathOptInterface.ScalarConstantChange — Type.

ScalarConstantChange{T}(new_constant)

A struct used to request a change in the constant term of a scalar-valued function. Applicable to ScalarAffineFunction and ScalarQuadraticFunction.

source

MathOptInterface.VectorConstantChange — Type.

VectorConstantChange{T}(new_constant)

A struct used to request a change in the constant vector of a vector-valued function. Applicable to VectorAffineFunction and VectorQuadraticFunction.

source

MathOptInterface.ScalarCoefficientChange — Type.

ScalarCoefficientChange{T}(variable, new_coefficient)

A struct used to request a change in the linear coefficient of a single variable in a scalar-valued function. Applicable to ScalarAffineFunction and ScalarQuadraticFunction.

source

MathOptInterface.MultirowChange — Type.

MultirowChange{T}(variable, rows, new_coefficients)

A struct used to request a change in the linear coefficients of a single variable in a vector-valued function. Applicable to VectorAffineFunction and VectorQuadraticFunction.

source

Sets

List of recognized sets.

MathOptInterface.AbstractSet — Type.

AbstractSet

Abstract supertype for set objects used to encode constraints.

source

MathOptInterface.Reals — Type.

Reals(dimension)

The set $\mathbb{R}^{dimension}$ (containing all points) of dimension dimension.

source

MathOptInterface.Zeros — Type.

Zeros(dimension)

The set $\{ 0 \}^{dimension}$ (containing only the origin) of dimension dimension.

source

MathOptInterface.Nonnegatives — Type.

Nonnegatives(dimension)

The nonnegative orthant $\{ x \in \mathbb{R}^{dimension} : x \ge 0 \}$ of dimension dimension.

source

MathOptInterface.Nonpositives — Type.

Nonpositives(dimension)

The nonpositive orthant $\{ x \in \mathbb{R}^{dimension} : x \le 0 \}$ of dimension dimension.

source

MathOptInterface.GreaterThan — Type.

GreaterThan{T <: Real}(lower::T)

The set $[lower,\infty) \subseteq \mathbb{R}$.

source

MathOptInterface.LessThan — Type.

LessThan{T <: Real}(upper::T)

The set $(-\infty,upper] \subseteq \mathbb{R}$.

source

MathOptInterface.EqualTo — Type.

EqualTo{T <: Real}(value::T)

The set containing the single point $x \in \mathbb{R}$ where $x$ is given by value.

source

MathOptInterface.Interval — Type.

Interval{T <: Real}(lower::T,upper::T)

The interval $[lower, upper] \subseteq \mathbb{R}$. If lower or upper is -Inf or Inf, respectively, the set is interpreted as a one-sided interval.

source

MathOptInterface.SecondOrderCone — Type.

SecondOrderCone(dimension)

The second-order cone (or Lorenz cone) $\{ (t,x) \in \mathbb{R}^{dimension} : t \ge || x ||_2 \}$ of dimension dimension.

source

MathOptInterface.RotatedSecondOrderCone — Type.

RotatedSecondOrderCone(dimension)

The rotated second-order cone $\{ (t,u,x) \in \mathbb{R}^{dimension} : 2tu \ge || x ||_2^2, t,u \ge 0 \}$ of dimension dimension.

source

MathOptInterface.GeometricMeanCone — Type.

GeometricMeanCone(dimension)

The geometric mean cone $\{ (t,x) \in \mathbb{R}^{n+1} : x \ge 0, t \le \sqrt[n]{x_1 x_2 \cdots x_n} \}$ of dimension dimension${}=n+1$.

source

MathOptInterface.ExponentialCone — Type.

ExponentialCone()

The 3-dimensional exponential cone $\{ (x,y,z) \in \mathbb{R}^3 : y \exp (x/y) \le z, y > 0 \}$.

source

MathOptInterface.DualExponentialCone — Type.

DualExponentialCone()

The 3-dimensional dual exponential cone $\{ (u,v,w) \in \mathbb{R}^3 : -u \exp (v/u) \le \exp(1) w, u < 0 \}$.

source

MathOptInterface.PowerCone — Type.

PowerCone{T <: Real}(exponent::T)

The 3-dimensional power cone $\{ (x,y,z) \in \mathbb{R}^3 : x^{exponent} y^{1-exponent} >= |z|, x \ge 0, y \ge 0 \}$ with parameter exponent.

source

MathOptInterface.DualPowerCone — Type.

DualPowerCone{T <: Real}(exponent::T)

The 3-dimensional power cone $\{ (u,v,w) \in \mathbb{R}^3 : (\frac{u}{exponent})^{exponent} (\frac{v}{1-exponent})^{1-exponent} \ge |w|, u \ge 0, v \ge 0 \}$ with parameter exponent.

source

MathOptInterface.PositiveSemidefiniteConeTriangle — Type.

PositiveSemidefiniteConeTriangle(dimension)

The (vectorized) cone of symmetric positive semidefinite matrices, with off-diagonals unscaled. The entries of the upper triangular part of the matrix are given column by column (or equivalently, the entries of the lower triangular part are given row by row). An $n \times n$ matrix has $n(n+1)/2$ lower-triangular elements, so for the vectorized cone of dimension $d$, the corresponding symmetric matrix has side dimension $\sqrt{1/4 + 2 d} - 1/2$ elements.

Examples

The matrix

\[\begin{bmatrix} 1 & 2 & 4\\ 2 & 3 & 5\\ 4 & 5 & 6 \end{bmatrix}\]

corresponds to $(1, 2, 3, 4, 5, 6)$ for PositiveSemidefiniteConeTriangle

Note

Two packed storage formats exist for symmetric matrices, the respective orders of the entries are:

upper triangular column by column (or lower triangular row by row);

lower triangular column by column (or upper triangular row by row).

The advantage of the first format is the mapping between the (i, j) matrix indices and the k index of the vectorized form. It is simpler and does not depend on the dimension of the matrix. Indeed,

the entry of matrix indices (i, j) has vectorized index k = div((j-1)*j, 2) + i if $i \leq j$ and k = div((i-1)*i, 2) + j if $j \leq i$;

and the entry with vectorized index k has matrix indices i = isqrt(2k) and j = k - div((i-1)*i, 2) or j = isqrt(2k) and i = k - div((j-1)*j, 2).

Duality note

The scalar product for the symmetric matrix in its vectorized form is the sum of the pairwise product of the diagonal entries plus twice the sum of the pairwise product of the upper diagonal entries; see [p. 634, 1]. This has important consequence for duality. Consider for example the following problem

\[\begin{align*} & \max_{x \in \mathbb{R}} & x \\ & \;\;\text{s.t.} & (1, -x, 1) & \in \text{PositiveSemidefiniteConeTriangle}(2). \end{align*}\]

The dual is the following problem

\[\begin{align*} & \min_{x \in \mathbb{R}^3} & y_1 + y_3 \\ & \;\;\text{s.t.} & 2y_2 & = 1\\ & & y & \in \text{PositiveSemidefiniteConeTriangle}(2). \end{align*}\]

Why do we use $2y_2$ in the dual constraint instead of $y_2$ ? The reason is that $2y_2$ is the scalar product between $y$ and the symmetric matrix whose vectorized form is $(0, 1, 0)$. Indeed, with our modified scalar products we have

\[\langle (0, 1, 0), (y_1, y_2, y_3) \rangle = \mathrm{trace} \begin{pmatrix} 0 & 1\\ 1 & 0 \end{pmatrix} \begin{pmatrix} y_1 & y_2\\ y_2 & y_3 \end{pmatrix} = 2y_2.\]

References

[1] Boyd, S. and Vandenberghe, L.. Convex optimization. Cambridge university press, 2004.

source

MathOptInterface.PositiveSemidefiniteConeSquare — Type.

PositiveSemidefiniteConeSquare(dimension)

The cone of symmetric positive semidefinite matrices. The entries of the matrix are given column by column (or equivalently, row by row). The matrix is both constrained to be symmetric and and to be positive semidefinite. That is, if the functions in entries $(i, j)$ and $(j, i)$ are different, then a constraint will be added to make sure that the entries are equal.

Examples

Constraining the matrix

\[\begin{bmatrix} 1 & -y\\ -z & 0\\ \end{bmatrix}\]

to be symmetric positive semidefinite can be achieved by constraining the vector $(1, -z, -y, 0)$ (or $(1, -y, -z, 0)$) to belong to the PositiveSemidefiniteConeSquare(2). It both constrains $y = z$ and $(1, -y, 0)$ (or $(1, -z, 0)$) to be in PositiveSemidefiniteConeTriangle.

source

MathOptInterface.LogDetConeTriangle — Type.

LogDetConeTriangle(dimension)

The Log-Determinant cone $\{ (t, X) \in \mathbb{R}^{1 + d(d+1)/2} : t \le \log(\det(X)) \}$ where the matrix X is represented in the same symmetric packed format as in the PositiveSemidefiniteConeTriangle. The argument dimension is the dimension of the matrix X, i.e., its number of rows or columns.

source

MathOptInterface.LogDetConeSquare — Type.

LogDetConeSquare(dimension)

The Log-Determinant cone $\{ (t, X) \in \mathbb{R}^{1 + d^2} : t \le \log(\det(X)), X \text{ symmetric} \}$ where the matrix X is represented in the same format as in the PositiveSemidefiniteConeSquare. Similarly to PositiveSemidefiniteConeSquare, constraints are added to ensures that X is symmetric. The argument dimension is the dimension of the matrix X, i.e., its number of rows or columns.

source

MathOptInterface.RootDetConeTriangle — Type.

RootDetConeTriangle(dimension)

The Root-Determinant cone $\{ (t, X) \in \mathbb{R}^{1 + d(d+1)/2} : t \le \det(X)^{1/d} \}$ where the matrix X is represented in the same symmetric packed format as in the PositiveSemidefiniteConeTriangle. The argument dimension is the dimension of the matrix X, i.e., its number of rows or columns.

source

MathOptInterface.RootDetConeSquare — Type.

RootDetConeSquare(dimension)

The Root-Determinant cone $\{ (t, X) \in \mathbb{R}^{1 + d^2} : t \le \det(X)^{1/d}, X \text{ symmetric} \}$ where the matrix X is represented in the same format as in the PositiveSemidefiniteConeSquare. Similarly to PositiveSemidefiniteConeSquare, constraints are added to ensures that X is symmetric. The argument dimension is the dimension of the matrix X, i.e., its number of rows or columns.

source

MathOptInterface.Integer — Type.

Integer()

The set of integers $\mathbb{Z}$.

source

MathOptInterface.ZeroOne — Type.

ZeroOne()

The set $\{ 0, 1 \}$.

source

MathOptInterface.Semicontinuous — Type.

Semicontinuous{T <: Real}(lower::T,upper::T)

The set $\{0\} \cup [lower,upper]$.

source

MathOptInterface.Semiinteger — Type.

Semiinteger{T <: Real}(lower::T,upper::T)

The set $\{0\} \cup \{lower,lower+1,\ldots,upper-1,upper\}$.

source

MathOptInterface.SOS1 — Type.

SOS1{T <: Real}(weights::Vector{T})

The set corresponding to the special ordered set (SOS) constraint of type 1. Of the variables in the set, at most one can be nonzero. The weights induce an ordering of the variables; as such, they should be unique values. The kth element in the set corresponds to the kth weight in weights. See here for a description of SOS constraints and their potential uses.

source

MathOptInterface.SOS2 — Type.

SOS2{T <: Real}(weights::Vector{T})

The set corresponding to the special ordered set (SOS) constraint of type 2. Of the variables in the set, at most two can be nonzero, and if two are nonzero, they must be adjacent in the ordering of the set. The weights induce an ordering of the variables; as such, they should be unique values. The kth element in the set corresponds to the kth weight in weights. See here for a description of SOS constraints and their potential uses.

source

Functions for getting and setting properties of sets.

MathOptInterface.dimension — Function.

dimension(s::AbstractVectorSet)

Return the underlying dimension (number of vector components) in the set s, i.e., $n$ if the set is a subset of $\mathbb{R}^n$.

source

Objective modifications

Functions for modifying objective functions. Use ObjectiveFunction and ObjectiveSense to set and query the objective function.

MathOptInterface.modifyobjective! — Function.

modifyobjective!(instance::AbstractInstance, change::AbstractFunctionModification)

Apply the modification specified by change to the objective function of instance. To change the function completely, call setobjective! instead.

Examples

modifyobjective!(instance, ScalarConstantChange(10.0))

source

MathOptInterface.canmodifyobjective — Function.

canmodifyobjective(instance::AbstractInstance, ::Type{M})::Bool where M<:AbstractFunctionModification

Return a Bool indicating whether it is possible to apply a modification of type M to the objective function of instance instance.

Examples

canmodifyobjective(instance, ScalarConstantChange{Float64})

source