Reference

API Reference

[Some introduction to API. List basic standalone methods.]

Attributes

List of attribute categories.

MathOptInterface.AbstractSolverAttribute — Type.

AbstractSolverAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of the solver.

source

MathOptInterface.AbstractSolverInstanceAttribute — Type.

AbstractSolverInstanceAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of the solver instance.

source

MathOptInterface.AbstractVariableAttribute — Type.

AbstractVariableAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of variables in the solver instance.

source

MathOptInterface.AbstractConstraintAttribute — Type.

AbstractConstraintAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of constraints in the solver instance.

source

Functions for getting and setting attributes.

MathOptInterface.cangetattribute — Function.

cangetattribute(s::AbstractSolver, attr::AbstractSolverAttribute)::Bool

Return a Bool indicating whether it is possible to query attribute attr from the solver s.

cangetattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::VariableReference)::Bool
cangetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::ConstraintReference{F,S})::Bool

Return a Bool indicating whether the solver instance m currently has a value for the attributed specified by attribute type attr applied to the variable reference v or constraint reference c.

cangetattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::Vector{VariableReference})::Bool
cangetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{ConstraintReference{F,S}})::Bool

Return a Bool indicating whether the solver instance m currently has a value for the attributed specified by attribute type attr applied to every variable references in v or constraint reference in c.

Examples

cangetattribute(m, ObjectiveValue())
cangetattribute(m, VariablePrimalStart(), varref)
cangetattribute(m, ConstraintPrimal(), conref)
cangetattribute(m, VariablePrimal(), [ref1, ref2])

source

MathOptInterface.getattribute — Function.

getattribute(s::AbstractSolver, attr::AbstractSolverAttribute)

Return an attribute attr of the solver s.

getattribute(m::AbstractSolverInstance, attr::AbstractSolverInstanceAttribute)

Return an attribute attr of the solver instance m.

getattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::VariableReference)

Return an attribute attr of the variable v in solver instance m.

getattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::Vector{VariableReference})

Return a vector of attributes corresponding to each variable in the collection v in the solver instance m.

getattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::ConstraintReference)

Return an attribute attr of the constraint c in solver instance m.

getattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{ConstraintReference{F,S}})

Return a vector of attributes corresponding to each constraint in the collection c in the solver instance m.

Examples

getattribute(m, ObjectiveValue())
getattribute(m, VariablePrimal(), ref)
getattribute(m, VariablePrimal(5), [ref1, ref2])
getattribute(m, OtherAttribute("something specific to cplex"))

source

MathOptInterface.getattribute! — Function.

getattribute!(output, m::AbstractSolverInstance, args...)

An in-place version of getattribute. The signature matches that of getattribute except that the the result is placed in the vector output.

source

MathOptInterface.cansetattribute — Function.

cansetattribute(s::AbstractSolver, attr::AbstractSolverAttribute)::Bool

Return a Bool indicating whether it is possible to set attribute attr in the solver s.

cansetattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, R::Type{VariableReference})::Bool
cangetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, R::Type{ConstraintReference{F,S})::Bool

Return a Bool indicating whether it is possible to set attribute attr applied to the reference type R in the solver instance m.

cansetattribute(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::Vector{VariableReference})::Bool
cansetattribute(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{ConstraintReference{F,S}})::Bool

Return a Bool indicating whether it is possible to set attribute attrapplied to every variable reference in v or constraint reference in c in the solver instance m.

Examples

cansetattribute(m, ObjectiveValue())
cansetattribute(m, VariablePrimalStart(), VariableReference)
cansetattribute(m, ConstraintPrimal(), ConstraintReference{VectorAffineFunction{Float64},Nonnegatives})

source

MathOptInterface.setattribute! — Function.

setattribute!(s::AbstractSolver, attr::AbstractSolverAttribute, value)

Assign value to the attribute attr of the solver s.

setattribute!(m::AbstractSolverInstance, attr::AbstractSolverInstanceAttribute, value)

Assign value to the attribute attr of the solver instance m.

setattribute!(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::VariableReference, value)

Assign value to the attribute attr of variable v in solver instance m.

setattribute!(m::AbstractSolverInstance, attr::AbstractVariableAttribute, v::Vector{VariableReference}, vector_of_values)

Assign a value respectively to the attribute attr of each variable in the collection v in solver instance m.

setattribute!(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::ConstraintReference, value)

Assign a value to the attribute attr of constraint c in solver instance m.

setattribute!(m::AbstractSolverInstance, attr::AbstractConstraintAttribute, c::Vector{ConstraintReference{F,S}})

Assign a value respectively to the attribute attr of each constraint in the collection c in solver instance m.

source

Solver

MathOptInterface.AbstractSolver — Type.

AbstractSolver

Abstract supertype for "solver" objects. A solver is a lightweight object used for selecting solvers and parameters. It does not store any solver instance data.

source

MathOptInterface.supportsproblem — Function.

supportsproblem(s::AbstractSolver, objective_type::F, constraint_types::Vector)::Bool

Return true if the solver supports optimizing a problem with objective type F and constraints of the types specified by constraint_types which is a list of tuples (F,S) for F-in-S constraints. Return false if the solver does not support this problem class.

Examples

supportsproblem(s, ScalarAffineFunction{Float64},
    [(ScalarAffineFunction{Float64},Zeros),
    (ScalarAffineFunction{Float64},LessThan),
    (ScalarAffineFunction{Float64},GreaterThan)])

should be true for a linear programming solver s.

supportsproblem(s, ScalarQuadraticFunction{Float64},
    [(ScalarAffineFunction{Float64},Zeros),
    (ScalarAffineFunction{Float64},LessThan),
    (ScalarAffineFunction{Float64},GreaterThan)])

should be true for a quadratic programming solver s.

supportsproblem(s, ScalarAffineFunction{Float64},
    [(ScalarAffineFunction{Float64},Zeros),
    (ScalarAffineFunction{Float64},LessThan),
    (ScalarAffineFunction{Float64},GreaterThan),
    (SingleVariable,ZeroOne)])

should be true for a mixed-integer linear programming solver s.

supportsproblem(s, ScalarAffineFunction{Float64},
    [(ScalarAffineFunction{Float64},Zeros),
    (ScalarAffineFunction{Float64},LessThan),
    (ScalarAffineFunction{Float64},GreaterThan),
    (VectorAffineFunction{Float64},SecondOrderCone)])

should be true for a second-order cone solver s.

source

List of solver attributes

MathOptInterface.SupportsDuals — Type.

SupportsDuals()

A Bool indicating if the solver should be expected to return dual solutions when appropriate.

source

MathOptInterface.SupportsAddConstraintAfterSolve — Type.

SupportsAddConstraintAfterSolve()

A Bool indicating if the solver supports adding constraints after a solve. If false, then a new solver instance should be constructed instead.

source

MathOptInterface.SupportsDeleteConstraint — Type.

SupportsDeleteConstraint()

A Bool indicating if the solver supports deleting constraints from a solver instance.

source

MathOptInterface.SupportsDeleteVariable — Type.

SupportsDeleteVariable()

A Bool indicating if the solver supports deleting variables from a solver instance.

source

MathOptInterface.SupportsAddVariableAfterSolve — Type.

SupportsAddVariableAfterSolve()

A Bool indicating if the solver supports adding variables after a solve. In the context of linear programming, this is known as column generation.

source

MathOptInterface.SupportsConicThroughQuadratic — Type.

SupportsConicThroughQuadratic()

A Bool indicating if the solver interprets certain quadratic constraints as second-order cone constraints.

source

Solver Instance

MathOptInterface.AbstractSolverInstance — Type.

AbstractSolverInstance

Abstract supertype which represents a solver's in-memory representation of an optimization problem.

source

MathOptInterface.SolverInstance — Function.

SolverInstance(solver::AbstractSolver)

Create a solver instance from the given solver.

source

MathOptInterface.optimize! — Function.

optimize!(m::AbstractSolverInstance)

Start the solution procedure.

source

MathOptInterface.free! — Function.

free!(m::AbstractSolverInstance)

Release any resources and memory used by the solver instance. Note that the Julia garbage collector takes care of this automatically, but automatic collection cannot always be forced. This method is useful for more precise control of resources, especially in the case of commercial solvers with licensing restrictions on the number of concurrent runs. Users must discard the solver instance object after this method is invoked.

source

List of solver instance attributes

MathOptInterface.RawSolver — Type.

RawSolver()

An object that may be used to access a solver-specific API for this solver instance.

source

MathOptInterface.ObjectiveSense — Type.

ObjectiveSense()

The sense of the objective function, an OptimizationSense with value MinSense, MaxSense, or FeasiblitySense.

source

MathOptInterface.NumberOfVariables — Type.

NumberOfVariables()

The number of variables in the solver instance.

source

MathOptInterface.ListOfVariableReferences — Type.

ListOfVariableReferences()

A Vector{VariableReference} with references to all variables present in the solver instance (i.e., of length equal to the value of NumberOfVariables()).

source

MathOptInterface.ListOfConstraints — Type.

ListOfConstraints()

A list of tuples of the form (F,S), where F is a function type and S is a set type indicating that the attribute NumberOfConstraints{F,S}() has value greater than zero.

source

MathOptInterface.NumberOfConstraints — Type.

NumberOfConstraints{F,S}()

The number of constraints of the type F-in-S present in the solver instance.

source

MathOptInterface.ListOfConstraintReferences — Type.

ListOfConstraintReferences{F,S}()

A Vector{ConstraintReferences{F,S}} with references to all constraints of type F-inS in the solver instance (i.e., of length equal to the value of NumberOfConstraints{F,S}()).

source

MathOptInterface.ResultCount — Type.

ResultCount()

The number of results available.

source

MathOptInterface.ObjectiveFunction — Type.

ObjectiveFunction()

An AbstractFunction instance which represents the objective function. It is guaranteed to be equivalent but not necessarily identical to the function provided by the user.

source

MathOptInterface.ObjectiveValue — Type.

ObjectiveValue(resultidx::Int=1)

The objective value of the resultindexth primal result.

source

MathOptInterface.ObjectiveBound — Type.

ObjectiveBound()

The best known bound on the optimal objective value.

source

MathOptInterface.RelativeGap — Type.

RelativeGap()

The final relative optimality gap, defined as $\frac{|b-f|}{|f|}$, where $b$ is the best bound and $f$ is the best feasible objective value.

source

MathOptInterface.SolveTime — Type.

SolveTime()

The total elapsed solution time (in seconds) as reported by the solver.

source

MathOptInterface.SimplexIterations — Type.

SimplexIterations()

The cumulative number of simplex iterations during the optimization process. In particular, for a mixed-integer program (MIP), the total simplex iterations for all nodes.

source

MathOptInterface.BarrierIterations — Type.

BarrierIterations()

The cumulative number of barrier iterations while solving a problem.

source

MathOptInterface.NodeCount — Type.

NodeCount()

The total number of branch-and-bound nodes explored while solving a mixed-integer program (MIP).

source

MathOptInterface.TerminationStatus — Type.

TerminationStatus()

A TerminationStatusCode explaining why the solver stopped.

source

MathOptInterface.PrimalStatus — Type.

PrimalStatus(N)
PrimalStatus()

The ResultStatusCode of the primal result N. If N is omitted, it defaults to 1.

source

MathOptInterface.DualStatus — Type.

DualStatus(N)
DualStatus()

The ResultStatusCode of the dual result N. If N is omitted, it defaults to 1.

source

Termination Status

The TerminationStatus attribute indicates why the solver stopped executing. The value of the attribute is of type TerminationStatusCode.

MathOptInterface.TerminationStatusCode — Type.

TerminationStatusCode

An Enum of possible values for the TerminationStatus attribute. This attribute is meant to explain the reason why the solver stopped executing.

OK

These are generally OK statuses.

Success: the algorithm ran successfully and has a result; this includes cases where the algorithm converges to an infeasible point (NLP) or converges to a solution of a homogeneous self-dual problem and has a certificate of primal/dual infeasibility

InfeasibleNoResult: the algorithm stopped because it decided that the problem is infeasible but does not have a result to return

UnboundedNoResult: the algorithm stopped because it decided that the problem is unbounded but does not have a result to return

InfeasibleOrUnbounded: the algorithm stopped because it decided that the problem is infeasible or unbounded (no result is available); this occasionally happens during MIP presolve

Limits

The solver stopped because of some user-defined limit. To be documented: IterationLimit, TimeLimit, NodeLimit, SolutionLimit, MemoryLimit, ObjectiveLimit, NormLimit, OtherLimit.

Problematic

This group of statuses means that something unexpected or problematic happened.

SlowProgress: the algorithm stopped because it was unable to continue making progress towards the solution

AlmostSuccess should be used if there is additional information that relaxed convergence tolerances are satisfied

To be documented: NumericalError, InvalidInstance, InvalidOption, Interrupted, OtherError.

source

Result Status

The PrimalStatus and DualStatus attributes indicate how to interpret the result returned by the solver. The value of the attribute is of type ResultStatusCode.

MathOptInterface.ResultStatusCode — Type.

ResultStatusCode

An Enum of possible values for the PrimalStatus and DualStatus attributes. The values indicate how to interpret the result vector.

FeasiblePoint

NearlyFeasiblePoint

InfeasiblePoint

InfeasibilityCertificate

NearlyInfeasibilityCertificate

ReductionCertificate

NearlyReductionCertificate

UnknownResultStatus

OtherResultStatus

source

Variables and Constraints

Basis Status

The BasisStatus attribute of a variable or constraint describes its status with respect to a basis, if one is known. The value of the attribute is of type BasisStatusCode.

MathOptInterface.BasisStatusCode — Type.

BasisStatusCode

An Enum of possible values for the VariableBasisStatus and ConstraintBasisStatus attribute. This explains the status of a given element with respect to an optimal solution basis. Possible values are:

Basic: element is in the basis

Nonbasic: element is not in the basis

NonbasicAtLower: element is not in the basis and is at its lower bound

NonbasicAtUpper: element is not in the basis and is at its upper bound

SuperBasic: element is not in the basis but is also not at one of its bounds

source

References

VariableReference
ConstraintReference
candelete
isvalid
delete!(::AbstractSolverInstance,::AnyReference)
candelete

Variables

Functions for adding variables. For deleting, see references section.

MathOptInterface.addvariables! — Function.

addvariables!(m::AbstractSolverInstance, n::Int)::Vector{VariableReference}

Add n scalar variables to the solver instance, returning a vector of variable references.

source

MathOptInterface.addvariable! — Function.

addvariable!(m::AbstractSolverInstance)::VariableReference

Add a scalar variable to the solver instance, returning a variable reference.

source

List of attributes associated with variables. [category AbstractVariableAttribute] Calls to getattribute and setattribute! should include as an argument a single VariableReference or a vector of VariableReference objects.

MathOptInterface.VariablePrimalStart — Type.

VariablePrimalStart()

An initial assignment of the variables that the solver may use to warm-start the solve.

source

MathOptInterface.VariablePrimal — Type.

VariablePrimal(N)
VariablePrimal()

The assignment to the primal variables in result N. If N is omitted, it is 1 by default.

source

MathOptInterface.VariableBasisStatus — Type.

VariableBasisStatus()

Returns the BasisStatusCode of a given variable, with respect to an available optimal solution basis.

source

Constraints

Functions for adding and modifying constraints.

MathOptInterface.isvalid — Method.

isvalid(m::AbstractSolverInstance, ref::AnyReference)::Bool

Return a Bool indicating whether this reference refers to a valid object in the solver instance m.

source

MathOptInterface.addconstraint! — Function.

addconstraint!(m::AbstractSolverInstance, func::F, set::S)::ConstraintReference{F,S} where {F,S}

Add the constraint $f(x) \in \mathcal{S}$ where $f$ is defined by func, and $\mathcal{S}$ is defined by set.

addconstraint!(m::AbstractSolverInstance, v::VariableReference, set::S)::ConstraintReference{SingleVariable,S} where {S}
addconstraint!(m::AbstractSolverInstance, vec::Vector{VariableReference}, set::S)::ConstraintReference{VectorOfVariables,S} where {S}

Add the constraint $v \in \mathcal{S}$ where $v$ is the variable (or vector of variables) referenced by v and $\mathcal{S}$ is defined by set.

source

MathOptInterface.addconstraints! — Function.

addconstraints!(m::AbstractSolverInstance, funcs::Vector{F}, sets::Vector{S})::Vector{ConstraintReference{F,S}} where {F,S}

Add the set of constraints specified by each function-set pair in funcs and sets. F and S should be concrete types. This call is equivalent to addconstraint!.(m, funcs, sets) but may be more efficient.

source

MathOptInterface.modifyconstraint! — Function.

Modify Function

modifyconstraint!(m::AbstractSolverInstance, c::ConstraintReference{F,S}, func::F)

Replace the function in constraint c with func. F must match the original function type used to define the constraint.

Examples

If c is a ConstraintReference{ScalarAffineFunction,S} and v1 and v2 are VariableReference objects,

modifyconstraint!(m, c, ScalarAffineFunction([v1,v2],[1.0,2.0],5.0))
modifyconstraint!(m, c, SingleVariable(v1)) # Error

Modify Set

modifyconstraint!(m::AbstractSolverInstance, c::ConstraintReference{F,S}, set::S)

Change the set of constraint c to the new set set which should be of the same type as the original set.

Examples

If c is a ConstraintReference{F,Interval}

modifyconstraint!(m, c, Interval(0, 5))
modifyconstraint!(m, c, NonPositives) # Error

Partial Modifications

modifyconstraint!(m::AbstractSolverInstance, c::ConstraintReference, change::AbstractFunctionModification)

Apply the modification specified by change to the function of constraint c.

Examples

modifyconstraint!(m, c, ScalarConstantChange(10.0))

source

MathOptInterface.canmodifyconstraint — Function.

Modify Function

canmodifyconstraint(m::AbstractSolverInstance, c::ConstraintReference{F,S}, func::F)::Bool

Return a Bool indicating whether it is possible to replace the function in constraint c with func. F must match the original function type used to define the constraint.

Examples

If c is a ConstraintReference{ScalarAffineFunction,S} and v1 and v2 are VariableReference objects,

canmodifyconstraint(m, c, ScalarAffineFunction([v1,v2],[1.0,2.0],5.0))
canmodifyconstraint(m, c, SingleVariable(v1)) # false

Modify Set

canmodifyconstraint(m::AbstractSolverInstance, c::ConstraintReference{F,S}, set::S)::Bool

Return a Bool indicating whether it is possible to change the set of constraint c to the new set set which should be of the same type as the original set.

Examples

If c is a ConstraintReference{F,Interval}

canmodifyconstraint(m, c, Interval(0, 5))
canmodifyconstraint(m, c, NonPositives) # false

Partial Modifications

canmodifyconstraint(m::AbstractSolverInstance, c::ConstraintReference, change::AbstractFunctionModification)::Bool

Return a Bool indicating whether it is possible to apply the modification specified by change to the function of constraint c.

Examples

canmodifyconstraint(m, c, ScalarConstantChange(10.0))

source

MathOptInterface.transformconstraint! — Function.

Transform Constraint Set

transformconstraint!(m::AbstractSolverInstance, c::ConstraintReference{F,S1}, newset::S2)::ConstraintReference{F,S2}

Replace the set in constraint c with newset. The constraint reference c will no longer be valid, and the function returns a new constraint reference.

Solvers may only support a subset of constraint transforms that they perform efficiently (for example, changing from a LessThan to GreaterThan set). In addition, set modification (where S1 = S2) should be performed via the modifyconstraint! function.

Typically, the user should delete the constraint and add a new one.

Examples

If c is a ConstraintReference{ScalarAffineFunction{Float64},LessThan{Float64}},

c2 = transformconstraint!(m, c, GreaterThan(0.0))
transformconstraint!(m, c, LessThan(0.0)) # errors

source

MathOptInterface.cantransformconstraint — Function.

Transform Constraint Set

cantransformconstraint(m::AbstractSolverInstance, c::ConstraintReference{F,S1}, newset::S2)::Bool

Return a Bool is the set in constraint c can be replaced with newset.

Examples

If c is a ConstraintReference{ScalarAffineFunction{Float64},LessThan{Float64}},

cantransformconstraint(m, c, GreaterThan(0.0)) # true
cantransformconstraint(m, c, ZeroOne())        # false

source

List of attributes associated with constraints. [category AbstractConstraintAttribute] Calls to getattribute and setattribute! should include as an argument a single ConstraintReference or a vector of ConstraintReference{F,S} objects.

MathOptInterface.ConstraintPrimalStart — Type.

ConstraintPrimalStart()

An initial assignment of the constraint primal values that the solver may use to warm-start the solve.

source

MathOptInterface.ConstraintDualStart — Type.

ConstraintDualStart()

An initial assignment of the constraint duals that the solver may use to warm-start the solve.

source

MathOptInterface.ConstraintPrimal — Type.

ConstraintPrimal(N)
ConstraintPrimal()

The assignment to the constraint primal values in result N. If N is omitted, it is 1 by default.

source

MathOptInterface.ConstraintDual — Type.

ConstraintDual(N)
ConstraintDual()

The assignment to the constraint dual values in result N. If N is omitted, it is 1 by default.

source

MathOptInterface.ConstraintBasisStatus — Type.

ConstraintBasisStatus()

Returns the BasisStatusCode of a given constraint, with respect to an available optimal solution basis.

source

MathOptInterface.ConstraintFunction — Type.

ConstraintFunction()

Return the AbstractFunction object used to define the constraint. It is guaranteed to be equivalent but not necessarily identical to the function provided by the user.

source

MathOptInterface.ConstraintSet — Type.

ConstraintSet()

Return the AbstractSet object used to define the constraint.

source

Functions and function modifications

List of recognized functions.

MathOptInterface.AbstractFunction — Type.

AbstractFunction

Abstract supertype for function objects.

source

MathOptInterface.SingleVariable — Type.

SingleVariable(variable)

The function that extracts the scalar variable referenced by variable, a VariableReference. This function is naturally be used for single variable bounds or integrality constraints.

source

MathOptInterface.VectorOfVariables — Type.

VectorOfVariables(variables)

The function that extracts the vector of variables referenced by variables, a Vector{VariableReference}. This function is naturally be used for constraints that apply to groups of variables, such as an "all different" constraint, an indicator constraint, or a complementarity constraint.

source

MathOptInterface.ScalarAffineFunction — Type.

ScalarAffineFunction{T}(variables, coefficients, constant)

The scalar-valued affine function $a^T x + b$, where:

$a$ is a sparse vector specified in tuple form by variables::Vector{VariableReference} and coefficients::Vector{T}

$b$ is a scalar specified by constant::T

Duplicate variable references in variables are accepted, and the corresponding coefficients are summed together.

source

MathOptInterface.VectorAffineFunction — Type.

VectorAffineFunction{T}(outputindex, variables, coefficients, constant)

The vector-valued affine function $A x + b$, where:

$A$ is a sparse matrix specified in triplet form by outputindex, variables, coefficients

$b$ is a vector specified by constant

Duplicate indices in the $A$ are accepted, and the corresponding coefficients are summed together.

source

MathOptInterface.ScalarQuadraticFunction — Type.

ScalarQuadraticFunction{T}(affine_variables, affine_coefficients, quadratic_rowvariables, quadratic_colvariables, quadratic_coefficients, constant)

The scalar-valued quadratic function $\frac{1}{2}x^TQx + a^T x + b$, where:

$a$ is a sparse vector specified in tuple form by affine_variables, affine_coefficients

$b$ is a scalar specified by constant

$Q$ is a symmetric matrix is specified in triplet form by quadratic_rowvariables, quadratic_colvariables, quadratic_coefficients

Duplicate indices in $a$ or $Q$ are accepted, and the corresponding coefficients are summed together. "Mirrored" indices (q,r) and (r,q) (where r and q are VariableReferences) are considered duplicates; only one need be specified.

source

MathOptInterface.VectorQuadraticFunction — Type.

VectorQuadraticFunction{T}(affine_variables, affine_coefficients, quadratic_rowvariables, quadratic_colvariables, quadratic_coefficients, constant)

The vector-valued quadratic function with ith component ("output index") defined as $\frac{1}{2}x^TQ_ix + a_i^T x + b_i$, where:

$a_i$ is a sparse vector specified in tuple form by the subset of affine_variables, affine_coefficients for the indices k where affine_outputindex[k] == i.

$b_i$ is a scalar specified by constant[i]

$Q_i$ is a symmetric matrix is specified in triplet form by the subset of quadratic_rowvariables, quadratic_colvariables, quadratic_coefficients for the indices k where quadratic_outputindex[k] == i

Duplicate indices in $a_i$ or $Q_i$ are accepted, and the corresponding coefficients are summed together. "Mirrored" indices (q,r) and (r,q) (where r and q are VariableReferences) are considered duplicates; only one need be specified.

source

List of function modifications.

MathOptInterface.ScalarConstantChange — Type.

ScalarConstantChange{T}(new_constant)

A struct used to request a change in the constant term of a scalar-valued function. Applicable to ScalarAffineFunction and ScalarQuadraticFunction.

source

MathOptInterface.VectorConstantChange — Type.

VectorConstantChange{T}(new_constant)

A struct used to request a change in the constant vector of a vector-valued function. Applicable to VectorAffineFunction and VectorQuadraticFunction.

source

MathOptInterface.ScalarCoefficientChange — Type.

ScalarCoefficientChange{T}(variable, new_coefficient)

A struct used to request a change in the linear coefficient of a single variable in a scalar-valued function. Applicable to ScalarAffineFunction and ScalarQuadraticFunction.

source

MathOptInterface.MultirowChange — Type.

MultirowChange{T}(variable, rows, new_coefficients)

A struct used to request a change in the linear coefficients of a single variable in a vector-valued function. Applicable to VectorAffineFunction and VectorQuadraticFunction.

source

Sets

List of recognized sets.

MathOptInterface.AbstractSet — Type.

AbstractSet

Abstract supertype for set objects used to encode constraints.

source

MathOptInterface.Reals — Type.

Reals(dim)

The set $\mathbb{R}^{dim}$ (containing all points) of dimension dim.

source

MathOptInterface.Zeros — Type.

Zeros(dim)

The set $\{ 0 \}^{dim}$ (containing only the origin) of dimension dim.

source

MathOptInterface.Nonnegatives — Type.

Nonnegatives(dim)

The nonnegative orthant $\{ x \in \mathbb{R}^{dim} : x \ge 0 \}$ of dimension dim.

source

MathOptInterface.Nonpositives — Type.

Nonpositives(dim)

The nonpositive orthant $\{ x \in \mathbb{R}^{dim} : x \le 0 \}$ of dimension dim.

source

MathOptInterface.GreaterThan — Type.

GreaterThan(lower)

The set $[lower,\infty) \subseteq \mathbb{R}$.

source

MathOptInterface.LessThan — Type.

LessThan(upper)

The set $(-\infty,upper] \subseteq \mathbb{R}$.

source

MathOptInterface.EqualTo — Type.

EqualTo(value)

The set containing the single point $x \in \mathbb{R}$ where $x$ is given by value.

source

MathOptInterface.Interval — Type.

Interval(lower,upper)

The interval $[lower, upper] \subseteq \mathbb{R}$. If lower or upper is -Inf or Inf, respectively, the set is interpreted as a one-sided interval.

source

MathOptInterface.SecondOrderCone — Type.

SecondOrderCone(dim)

The second-order cone (or Lorenz cone) $\{ (t,x) \in \mathbb{R}^{dim} : t \ge || x ||_2 \}$ of dimension dim.

source

MathOptInterface.RotatedSecondOrderCone — Type.

RotatedSecondOrderCone(dim)

The rotated second-order cone $\{ (t,u,x) \mathbb{R}^{dim} : 2tu \ge || x ||_2^2, t,u \ge 0 \}$ of dimension dim.

source

MathOptInterface.ExponentialCone — Type.

ExponentialCone()

The 3-dimensional exponential cone $\{ (x,y,z) \in \mathbb{R}^3 : y \exp (x/y) \le z, y > 0 \}$.

source

MathOptInterface.DualExponentialCone — Type.

DualExponentialCone()

The 3-dimensional dual exponential cone $\{ (u,v,w) \in \mathbb{R}^3 : -u \exp (v/u) \le \exp(1) w, u < 0 \}$.

source

MathOptInterface.PowerCone — Type.

PowerCone(a)

The 3-dimensional power cone $\{ (x,y,z) \in \mathbb{R}^3 : x^{a} y^{1-a} >= |z|, x \ge 0, y \ge 0 \}$ with parameter a.

source

MathOptInterface.DualPowerCone — Type.

DualPowerCone(a)

The 3-dimensional power cone $\{ (u,v,w) \in \mathbb{R}^3 : (\frac{u}{a})^a (\frac{v}/{1-a})^{1-a} >= |w|, u \ge 0, v \ge 0 \}$ with parameter a.

source

MathOptInterface.PositiveSemidefiniteConeTriangle — Type.

PositiveSemidefiniteConeTriangle(dim)

The (vectorized) cone of symmetric positive semidefinite matrices, with off-diagonals unscaled. The entries of the upper triangular part of the matrix are given row by row (or equivalently, the entries of the lower triangular part are given column by column). An $n \times n$ matrix has $n(n+1)/2$ lower-triangular elements, so for the vectorized cone of dimension dim, the corresponding symmetric matrix has side dimension $\sqrt (1/4 + 2 dim) - 1/2$ elements. The scalar product is the sum of the pairwise product of the diagonal entries plus twice the sum of the pairwise product of the upper diagonal entries.

Examples

The matrix

\[\begin{bmatrix} 1 & 2 & 3\\ 2 & 4 & 5\\ 3 & 5 & 6 \end{bmatrix}\]

corresponds to $(1, 2, 3, 4, 5, 6)$ for PositiveSemidefiniteConeTriangle

source

MathOptInterface.PositiveSemidefiniteConeScaled — Type.

PositiveSemidefiniteConeScaled(dim)

The (vectorized) cone of symmetric positive semidefinite matrices, with off-diagonals scaled. The entries of the upper triangular part of the matrix are given row by row (or equivalently, the entries of the lower triangular part are given column by column). An $n \times n$ matrix has $n(n+1)/2$ lower-triangular elements, so for the vectorized cone of dimension dim, the corresponding symmetric matrix has side dimension $\sqrt (1/4 + 2 dim) - 1/2$ elements. The off-diagonal entries of the matrices of both the cone and its dual are scaled by $\sqrt{2}$ and the scalar product is simply the sum of the pairwise product of the entries.

Examples

The matrix

\[\begin{bmatrix} 1 & 2 & 3\\ 2 & 4 & 5\\ 3 & 5 & 6 \end{bmatrix}\]

and to $(1, 2\sqrt{2}, 3\sqrt{2}, 4, 5\sqrt{2}, 6)$ for PositiveSemidefiniteConeScaled.

source

MathOptInterface.Integer — Type.

Integer()

The set of integers $\mathbb{Z}$.

source

MathOptInterface.ZeroOne — Type.

ZeroOne()

The set $\{ 0, 1 \}$.

source

MathOptInterface.Semicontinuous — Type.

Semicontinuous(l,u)

The set $\{0\} \cup [l,u]$.

source

MathOptInterface.Semiinteger — Type.

Semiinteger(l,u)

The set $\{0\} \cup \{l,l+1,\ldots,u-1,u\}$.

source

MathOptInterface.SOS1 — Type.

SOS1(weights)

The set corresponding to the special ordered set (SOS) constraint of type 1. Of the variables in the set, at most one can be nonzero. The weights induce an ordering of the variables; as such, they should be unique values. The kth element in the set corresponds to the kth weight in weights. See here for a description of SOS constraints and their potential uses.

source

MathOptInterface.SOS2 — Type.

SOS2(weights)

The set corresponding to the special ordered set (SOS) constraint of type 2. Of the variables in the set, at most two can be nonzero, and if two are nonzero, they must be adjacent in the ordering of the set. The weights induce an ordering of the variables; as such, they should be unique values. The kth element in the set corresponds to the kth weight in weights. See here for a description of SOS constraints and their potential uses.

source

Functions for getting and setting properties of sets.

MathOptInterface.dimension — Function.

dimension(s::AbstractVectorSet)

Return the underlying dimension (number of vector components) in the set s, i.e., $n$ if the set is a subset of $\mathbb{R}^n$.

source

Objectives

Functions for adding and modifying objectives.

setobjective!
modifyobjective!
canmodifyobjective